% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAccount_server.R
\name{accountDefinition}
\alias{accountDefinition}
\title{Account Definition Module}
\usage{
accountDefinition(input, output, session, dbSettings,
  active = reactive(TRUE))
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{active}{Reactive expression whether the module state should be updated.}
}
\value{
Empty list.
}
\description{
Server logic to define an account.
}
