% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{actionButton}
\alias{actionButton}
\title{Action button/link}
\usage{
actionButton(class = c("btn", "btn-primary", "btn-flamingo"), ...)
}
\arguments{
\item{class}{HTML class attribute}

\item{...}{arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}}
}
\description{
Modified version of the default \code{\link[shiny:actionButton]{shiny::actionButton()}}
}
