% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anainputs_module.R
\name{anainputsUI}
\alias{anainputsUI}
\alias{anainputs}
\title{anainputsUI}
\usage{
anainputsUI(id, heading = NULL, collapsible = FALSE, show = FALSE)

anainputs(input, output, session, analysisID, refresh_opt = TRUE,
  counter = reactive(NULL), active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{heading}{Panel title.}

\item{collapsible}{Option to collapse panel, Default FALSE.}

\item{show}{Show panel. Default FALSE.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis ID.}

\item{refresh_opt}{Option to hide/show refresh button.}

\item{counter}{Reactive value storing actionButton status.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.
}
\description{
UI/View for inputs of an analysis.

Server logic for inputs of an analysis.
}
