% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_details_module.R
\name{analysis_detailsUI}
\alias{analysis_detailsUI}
\alias{analysis_details}
\title{analysis_detailsUI}
\usage{
analysis_detailsUI(id)

analysis_details(input, output, session, analysisID, portfolioID, counter)
}
\arguments{
\item{id}{module id}

\item{analysisID}{Selected analysis ID.}

\item{portfolioID}{Selected portfolio ID.}

\item{id}{module id}

\item{reload_generated}{Imports function to reload Generated Inputs table.}

\item{id}{module id}
}
\description{
UI side of function wrapping panel to show analyses details table.

Server side of function wrapping panel to show analyses details table.
}
