% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{api_delete_analyses_id}
\alias{api_delete_analyses_id}
\title{Delete analyses id}
\usage{
api_delete_analyses_id(id)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}
}
\value{
Response to analysis deletion.
}
\description{
Removes an analysis.
}
