% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{api_delete_portfolios_id}
\alias{api_delete_portfolios_id}
\title{Delete portfolios by id}
\usage{
api_delete_portfolios_id(id)
}
\arguments{
\item{id}{a unique integer value identifying this portfolio.}
}
\description{
Delete portfolios by id
}
