% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{api_get_analyses_id}
\alias{api_get_analyses_id}
\title{Get analyses id}
\usage{
api_get_analyses_id(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted analyses id.
}
\description{
Returns the specific analysis entry.
}
