% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_get_analyses_input_errors_file}
\alias{api_get_analyses_input_errors_file}
\title{Get analysis input_errors file}
\usage{
api_get_analyses_input_errors_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted analysis input_errors file.
}
\description{
Gets the analysis input_errors_file contents
}
