% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_get_analyses_input_file}
\alias{api_get_analyses_input_file}
\title{Get analysis input file}
\usage{
api_get_analyses_input_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted analysis input file.
}
\description{
Gets the analysis input_file contents
}
