% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{api_get_analyses_inputs_tar}
\alias{api_get_analyses_inputs_tar}
\title{Get analysis input tar}
\usage{
api_get_analyses_inputs_tar(id, dest = tempfile(fileext = ".tar"))
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{dest}{path where to download tar file.}
}
\value{
API respone.
}
\description{
Downloads the analysis inputs tar.
}
