% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{api_get_analyses_outputs_tar}
\alias{api_get_analyses_outputs_tar}
\title{Get analysis outputs tar}
\usage{
api_get_analyses_outputs_tar(id, dest = tempfile(fileext = ".tar"))
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{dest}{path to download tar file}
}
\value{
Previously posted analysis outputs tar file.
}
\description{
Downloads the analysis outputs tar.
}
