% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_get_analyses_settings_file}
\alias{api_get_analyses_settings_file}
\title{Get analysis settings file}
\usage{
api_get_analyses_settings_file(id)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}
}
\value{
Previously posted analysis settings file.
}
\description{
Gets the analysis settings_file contents.
}
