% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{api_get_healthcheck}
\alias{api_get_healthcheck}
\title{Perform healthcheck}
\usage{
api_get_healthcheck()
}
\value{
response containing status of API connection
}
\description{
Gets the current status of the api.
}
