% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{api_get_healthcheck}
\alias{api_get_healthcheck}
\title{Perform healthcheck}
\usage{
api_get_healthcheck()
}
\arguments{
\item{...}{Other inputs.}
}
\value{
Helthcheck files.
}
\description{
Gets the current status of the api.
}
\details{
Typo in he(a)lthcheck reflects typo in the API.
}
