% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{api_get_models}
\alias{api_get_models}
\title{Get models}
\usage{
api_get_models(supplier_id = "")
}
\arguments{
\item{supplier_id}{the supplier ID for the model. Default is empty string.}
}
\value{
previously posted models. . Default empty string returns all portfolios.
}
\description{
Returns a list of model objects.
}
