% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models_id.R
\name{api_get_models_id_resource_file}
\alias{api_get_models_id_resource_file}
\title{Get model resource file}
\usage{
api_get_models_id_resource_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this model.}
}
\value{
previously posted model resource file.
}
\description{
Gets the model resource_file contents
}
