% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_get_portfolios_accounts_file}
\alias{api_get_portfolios_accounts_file}
\title{Get portfolios accounts file}
\usage{
api_get_portfolios_accounts_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this portfolio.}
}
\value{
the previously posted portfolio accounts file.
}
\description{
Gets the portfolios accounts_file contents
}
