% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolio.R
\name{api_get_portfolios_id}
\alias{api_get_portfolios_id}
\title{Get portfolios id}
\usage{
api_get_portfolios_id(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted portfolios id.
}
\description{
Returns the specific analysis entry.
}
