% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolio.R
\name{api_get_portfolios_location_file}
\alias{api_get_portfolios_location_file}
\title{Get portfolios location file}
\usage{
api_get_portfolios_location_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted portfolios location files.
}
\description{
Gets the portfolios location_file contents
}
