% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_get_portfolios_reinsurance_info_file}
\alias{api_get_portfolios_reinsurance_info_file}
\title{Get portfolios reinsurance info file}
\usage{
api_get_portfolios_reinsurance_info_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this portfolio.}
}
\value{
the previously posted portfolio reinsurance info file.
}
\description{
Gets the portfolios reinsurance_info_file contents
}
