% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_get_portfolios_reinsurance_scope_file}
\alias{api_get_portfolios_reinsurance_scope_file}
\title{Get portfolios reinsurance source file}
\usage{
api_get_portfolios_reinsurance_scope_file(id)
}
\arguments{
\item{id}{A unique integer value identifying this portfolio.}
}
\value{
The previously posted portfolio reinsurance source file.
}
\description{
Gets the portfolios reinsurance_scope_file contents.
}
