% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{api_init}
\alias{api_init}
\title{API initialization}
\usage{
api_init(host, port, scheme = c("http", "https"))
}
\arguments{
\item{port}{host port}

\item{scheme}{communication scheme}
}
\description{
Builds API URL.
}
