% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{api_post_analyses}
\alias{api_post_analyses}
\title{Post analyses}
\usage{
api_post_analyses(name, portfolio, model)
}
\arguments{
\item{name}{The name of the analysis.}

\item{portfolio}{The id of the portfolio.}

\item{model}{The id of the model.}
}
\value{
The posted analysis.
}
\description{
Creates an analysis based on the input data.
}
