% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_cancel}
\alias{api_post_analyses_cancel}
\title{Cancel analysis}
\usage{
api_post_analyses_cancel(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
the cancelled analysis.
}
\description{
Cancel the selected analysis.
}
