% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_cancel_generate_inputs}
\alias{api_post_analyses_cancel_generate_inputs}
\title{Cancel analysis input generation}
\usage{
api_post_analyses_cancel_generate_inputs(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
the cancelled analysis input generation.
}
\description{
Cancel the inputs generated in the analysis.
}
