% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_input_errors_file}
\alias{api_post_analyses_input_errors_file}
\title{Post analysis input_errors file}
\usage{
api_post_analyses_input_errors_file(id, filepath_input_errors)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{filepath_input_errors}{Path to the input_errors file.}
}
\value{
The posted analysis input_errors file.
}
\description{
Sets the analysis input_errors_file contents.
}
