% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_input_file}
\alias{api_post_analyses_input_file}
\title{Post analysis input file}
\usage{
api_post_analyses_input_file(id, filepath_input)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_input}{path to the input file.}
}
\value{
the posted analysis input file.
}
\description{
Sets the analysis input_file contents.
}
