% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_input_generation_traceback_file}
\alias{api_post_analyses_input_generation_traceback_file}
\title{Post analysis input_generation_traceback file}
\usage{
api_post_analyses_input_generation_traceback_file(id,
  filepath_input_generation_traceback)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_input_generation_traceback}{path to the input_generation_traceback file.}
}
\value{
the posted analysis input_generation_traceback file.
}
\description{
Sets the analysis input_generation_traceback_file contents.
}
