% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_output_file}
\alias{api_post_analyses_output_file}
\title{Post analysis output file}
\usage{
api_post_analyses_output_file(id, filepath_output)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_output}{path to the output file.}
}
\value{
the posted analysis output file.
}
\description{
Sets the analysis output_file contents.
}
