% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_run}
\alias{api_post_analyses_run}
\title{Run analyses id}
\usage{
api_post_analyses_run(id)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}
}
\value{
Analysis status.
}
\description{
Returns the analysis status.
}
