% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_settings_file}
\alias{api_post_analyses_settings_file}
\title{Post analysis settings file}
\usage{
api_post_analyses_settings_file(id, filepath_settings)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_settings}{path to the settings file.}
}
\value{
the posted analysis settings file.
}
\description{
Sets the analysis settings_file contents.
}
