% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{api_post_portfolios}
\alias{api_post_portfolios}
\title{Post portfolios}
\usage{
api_post_portfolios(name)
}
\arguments{
\item{name}{Name of the portfolio.}
}
\value{
The posted portfolio.
}
\description{
Creates a portfolio based on the input data.
}
