% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{api_post_portfolios_create_analysis}
\alias{api_post_portfolios_create_analysis}
\title{Post portfolios create analysis}
\usage{
api_post_portfolios_create_analysis(id, name, model)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{name}{name of the portfolio.}
}
\value{
the posted portfolio analysis created.
}
\description{
Creates an analysis object from the portfolio.
}
