% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_post_portfolios_reinsurance_info_file}
\alias{api_post_portfolios_reinsurance_info_file}
\title{Post portfolios reinsurance info file}
\usage{
api_post_portfolios_reinsurance_info_file(id, filepath_reinsurance_info)
}
\arguments{
\item{id}{A unique integer value identifying this portfolio.}

\item{filepath_reinsurance_info}{Path to reinsurance info file.}
}
\value{
The posted portfolio reinsurance info file.
}
\description{
Sets the portfolios reinsurance_info contents.
}
