% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_post_portfolios_reinsurance_scope_file}
\alias{api_post_portfolios_reinsurance_scope_file}
\title{Post portfolios reinsurance source file}
\usage{
api_post_portfolios_reinsurance_scope_file(id,
  filepath_reinsurance_scope)
}
\arguments{
\item{id}{A unique integer value identifying this portfolio.}

\item{filepath_reinsurance_scope}{Path to reinsurance source file.}
}
\value{
the posted portfolio reinsurance source file.
}
\description{
Sets the portfolios reinsurance_scope contents.
}
