% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{api_put_portfolios_id}
\alias{api_put_portfolios_id}
\title{Put portfolios id}
\usage{
api_put_portfolios_id(id, name)
}
\arguments{
\item{id}{a unique integer value identifying this portfolio.}

\item{name}{name of the portfolio.}
}
\value{
the updated analysis.
}
\description{
Updates the specified analysis.
}
