% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{APIgetenv}
\alias{APIgetenv}
\title{flamingoAPI}
\usage{
APIgetenv(server, port, version, share_filepath)
}
\arguments{
\item{server}{API Host name.}

\item{port}{API Host port.}

\item{version}{API version.}

\item{share_filepath}{path of shared location where to download files.}
}
\value{
A settings object (list).
}
\description{
Creates am api settings object which can then be used
to create new connections to the Flamingo API
}
