% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apihub.R
\docType{class}
\name{APIHub}
\alias{APIHub}
\title{APIHub}
\format{\code{\link{R6Class}} object.}
\usage{
APIHub
}
\value{
Object of \code{\link{R6Class}} with methods for connecting to Oasis API.
}
\description{
R6 Class for OasisUI API Hub.
}
\section{Arguments}{

\describe{
\item{user}{string for username.}
\item{pwd}{string for user password.}
}
}

\section{Privates}{

\describe{
\item{yyyy}{yyyy.}
}
}

\section{Methods}{

\describe{

Initialize
 \item{\code{method(arg1,arg2)}}{dddd.}
}
}

\section{Usage}{

\preformatted{api_hub <- APIHub$new()
api_hub$method(arg)
}
}

\keyword{datasets}
