% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{associateModelUI}
\alias{associateModelUI}
\title{UI/View to define the parameters of a process run}
\usage{
associateModelUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the parameters of a process run
}
