% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{barPlot}
\alias{barPlot}
\title{barPlot}
\usage{
barPlot(xlabel, ylabel, titleToUse, data, multipleplots)
}
\arguments{
\item{xlabel}{label X axis}

\item{ylabel}{label y axis}

\item{titleToUse}{title}

\item{data}{dataframe to plot}

\item{multipleplots}{flag about facets}
}
\description{
bar plot
}
\details{
barPlot structure
}
