% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{basicplot}
\alias{basicplot}
\title{basicplot}
\usage{
basicplot(xlabel, ylabel, titleToUse, data)
}
\arguments{
\item{xlabel}{label x axis}

\item{ylabel}{label y axis}

\item{titleToUse}{plot title}

\item{data}{dataset to plot}
}
\description{
basic plot
}
\details{
basicplot structure
}
