% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAna_server.R, R/batchAna_ui.R
\name{batchAna}
\alias{batchAna}
\alias{batchAnaUI}
\title{batchAna}
\usage{
batchAna(input, output, session, dbSettings, apiSettings,
  active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

batchAnaUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{user}{reactive expression yielding user}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

List of tags.
}
\description{
Server logic to define a batch analysis.

UI/View to define a batch analysis.
}
