% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_server.R
\name{browseprogrammes}
\alias{browseprogrammes}
\title{browseprogrammes Module}
\usage{
browseprogrammes(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message, reloadMillis = 10000)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}
}
\value{
list of reactives:
\itemize{
		\item{\code{navigate}}{reactive yielding navigation}
		\item{\code{progOasisId}}{selected ProgOasis Id}
}
}
\description{
Server logic to define a programme
}
