% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_ui.R
\name{browseprogrammesUI}
\alias{browseprogrammesUI}
\title{browse_ui
Browse programmes UI}
\usage{
browseprogrammesUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
UI/View for the process run page
}
