% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{buildDbQuery}
\alias{buildDbQuery}
\title{Convenience function to construct query statements}
\usage{
buildDbQuery(dbFunc, ..., dboPrefix = TRUE, squareBrackets = FALSE)
}
\arguments{
\item{dbFunc}{database function to call}

\item{...}{arguments to \code{dbFunc}. \code{character()} args are quoted
with single quotes (').}

\item{dboPrefix}{prefix \code{dbFunc} with \code{"dbo."}}

\item{squareBrackets}{surround the argument list with [ and ]}
}
\description{
Constructs query statements that can be executed with
\link{dbExecute}.
}
