% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_utils.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize first letter of all words in a string}
\usage{
capitalize_first_letter(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
Returns a string with all words of input string have the first letter capitalized.
}
