% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loc.R
\name{check_loc}
\alias{check_loc}
\title{check_loc}
\usage{
check_loc(analysisID, portfolioID)
}
\arguments{
\item{analysisID}{Selected analysis id.}

\item{portfolioID}{selected portfolio ID.}
}
\description{
Compares uploaded and modeled locations.
}
