% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{collapseButton}
\alias{collapseButton}
\title{collapseButton}
\usage{
collapseButton(id, id_collapse, ..., width = NULL, collapsed = FALSE)
}
\arguments{
\item{id}{module id}

\item{id_collapse}{id of button that triggers collapsing.}

\item{width}{widget widt}

\item{collapsed}{FALSE.}
}
\value{
List of tags.
}
\description{
collapseButton
}
