% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompany_server.R
\name{companyDefinition}
\alias{companyDefinition}
\title{companyDefinition}
\usage{
companyDefinition(input, output, session, dbSettings, userId,
  active = reactive(TRUE), logMessage = message)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}
}
\value{
Empty list.
}
\description{
Server logic to define a company.
}
