% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompany_ui.R
\name{companyDefinitionUI}
\alias{companyDefinitionUI}
\title{companyDefinitionUI}
\usage{
companyDefinitionUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
List of tags.
}
\description{
UI/View to define an account.
}
