% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureAdvancedGUL}
\alias{configureAdvancedGUL}
\title{Function wrapping sub-panel to define outputs advanced configuration GUL}
\usage{
configureAdvancedGUL(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration GUL
}
