% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureAdvancedRI}
\alias{configureAdvancedRI}
\title{Function wrapping sub-panel to define outputs advanced configuration RI}
\usage{
configureAdvancedRI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration RI
}
