% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureOutputAdvancedGULUI}
\alias{configureOutputAdvancedGULUI}
\title{UI/View to define the Advanced  GUL parameters of a process run}
\usage{
configureOutputAdvancedGULUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  GUL parameters of a process run
}
