% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureOutputAdvancedILUI}
\alias{configureOutputAdvancedILUI}
\title{UI/View to define the Advanced  IL parameters of a process run}
\usage{
configureOutputAdvancedILUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  IL parameters of a process run
}
