% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureOutputAdvancedUI}
\alias{configureOutputAdvancedUI}
\title{UI/View to define the Advanced  Model parameters of a process run}
\usage{
configureOutputAdvancedUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  Model parameters of a process run
}
