% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{configureOutputPerilAdvancedUI}
\alias{configureOutputPerilAdvancedUI}
\title{UI/View to define the Advanced  Model perils of a process run}
\usage{
configureOutputPerilAdvancedUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  Model perils of a process run
}
