% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{construct_analysis_settings}
\alias{construct_analysis_settings}
\title{Construct analysis settings}
\usage{
construct_analysis_settings(inputsettings, outputsLossTypes)
}
\arguments{
\item{inputsettings}{List of input settings.}

\item{outputsLossTypes}{List of output configuration.}
}
\value{
List of analysis settings.
}
\description{
Constructs a list of analysis settings.
}
