% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{construct_analysis_settings}
\alias{construct_analysis_settings}
\title{Construct analysis settings list}
\usage{
construct_analysis_settings(inputsettings, outputsLossTypes)
}
\arguments{
\item{inputsettings}{list of input settings}

\item{outputsLossType}{list of output configuration}
}
\value{
list of analysis settings.
}
\description{
Constructs a list of analysis settings
}
\details{
Construct analysis_settingsList
}
