% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{convert_created_modified}
\alias{convert_created_modified}
\title{Convert created and modified columns}
\usage{
convert_created_modified(tbl_obj)
}
\arguments{
\item{tbl_obj}{Dataframe to convert.}
}
\value{
Dataframe with converted columns.
}
\description{
Replaces created and modified columns with different date format.
}
