% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{createFootprintMap}
\alias{createFootprintMap}
\title{Create Footprint Map}
\usage{
createFootprintMap(dbSettings, fileId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{fileId}{File id for exposure data.}
}
\description{
creates a footprint map using leaflet based on exposure data
stored in the flamingo database.
}
