% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHazardMap_module.R
\name{createHazardMapUI}
\alias{createHazardMapUI}
\alias{createHazardMap}
\title{createHazardMap}
\usage{
createHazardMapUI(id)

createHazardMap(input, output, session, file_map, file_pins)
}
\arguments{
\item{file_map}{File to plot as map.}

\item{file_pins}{File to plot pins.}
}
\value{
Leaflet map.
}
\description{
Creates a hazard map using leaflet.

Creates a hazard map using leaflet.
}
