% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{createPlainMap}
\alias{createPlainMap}
\title{Create Plain Map}
\usage{
createPlainMap(df)
}
\arguments{
\item{df}{df to plot as map}
}
\value{
Leaflet map.
}
\description{
Creates a plain map using leaflet.
}
