% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createProgOasis}
\alias{createProgOasis}
\title{Create Prog Oasis}
\usage{
createProgOasis(dbSettings, progId, modelId, transformId)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{progId}{programme id}

\item{modelId}{model id}

\item{transformId}{transform id}
}
\value{
Prog Oasis id
}
\description{
Create Prog Oasis
}
