% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineID.R
\name{defineIDUI}
\alias{defineIDUI}
\title{defineIDUI}
\usage{
defineIDUI(id, w, batch = FALSE)
}
\arguments{
\item{id}{module id}

\item{w}{width of the coulmn}

\item{i}{counter}
}
\description{
defineIDUI
}
