% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{defineProgramme}
\alias{defineProgramme}
\title{Function defining panel elements to define a programme}
\usage{
defineProgramme(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function defining panel elements to define a programme
}
