% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurevalidation_module.R
\name{exposurevalidationUI}
\alias{exposurevalidationUI}
\alias{exposurevalidation}
\title{exposurevalidationUI}
\usage{
exposurevalidationUI(id)

exposurevalidation(input, output, session, analysisID = "",
  portfolioID = "", counter = NULL, active = reactive(TRUE))
}
\value{
List of tags.
}
\description{
UI/View for exposure validation of an analysis.

Server logic for exposure validation of an analysis.
}
