% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurevalidation_map_module.R
\name{exposurevalidationmapUI}
\alias{exposurevalidationmapUI}
\alias{exposurevalidationmap}
\title{exposurevalidationmapUI}
\usage{
exposurevalidationmapUI(id)

exposurevalidationmap(input, output, session, analysisID = "",
  portfolioID = "", counter = NULL, active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis id.}

\item{portfolioID}{Selected portfolio ID.}

\item{counter}{Reactive value to trigger inputs download.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.
}
\description{
UI/View for exposure validation of an analysis.

Server logic for exposure validation of an analysis.
}
