% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{extract_analyses_output_file}
\alias{extract_analyses_output_file}
\title{Extract analyses output files}
\usage{
extract_analyses_output_file(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\description{
extract the output files in afixed location
}
