% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\name{extractColName}
\alias{extractColName}
\title{extractColName}
\usage{
extractColName(tbl_obj, col_type)
}
\arguments{
\item{tbl_obj}{object for which to retrieve column name.}

\item{col_type}{string of the column type to retrieve}
}
\description{
function to extract column name from named list including the column names
}
