% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileViewer_server.R
\docType{data}
\name{fileMgmtButtons}
\alias{fileMgmtButtons}
\title{File Management Buttons}
\format{An object of class \code{character} of length 6.}
\usage{
fileMgmtButtons
}
\description{
File Management Buttons
}
\keyword{datasets}
