% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileViewer_server.R
\docType{data}
\name{fileMgmtDivs}
\alias{fileMgmtDivs}
\title{File Management Divisions}
\format{An object of class \code{character} of length 6.}
\usage{
fileMgmtDivs
}
\description{
File Management Divisions
}
\keyword{datasets}
