% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileViewer_server.R, R/fileViewer_ui.R
\name{fileViewer}
\alias{fileViewer}
\alias{fileViewerUI}
\title{File Viewer Module}
\usage{
fileViewer(input, output, session, dbSettings,
  preselRunId = reactive(-1), active, logMessage = message)

fileViewerUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{preselRunId}{Reactive string expression for reselected run id from landingpage.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

LIst of tags.
}
\description{
Server logic to view files.

UI/View to view files.
}
