% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.construct_flamingo_from_hic.R
\name{flamingo.construct_flamingo_from_hic}
\alias{flamingo.construct_flamingo_from_hic}
\title{flamingo.construct_flamingo_from_hic}
\usage{
flamingo.construct_flamingo_from_hic(
  hic_file,
  normalization,
  resolution,
  chr_name,
  chr_size,
  alpha = -0.25
)
}
\arguments{
\item{hic_file}{Path to the hic file.}

\item{normalization}{Method of normalization. Must be one of "NONE", "VC", "VC_SQRT", "KR".}

\item{resolution}{Target resolution.}

\item{chr_name}{chromosome name}

\item{chr_size}{chromsome size}

\item{alpha}{convertion factor between interaction frequency and distance, default is -0.25}
}
\value{
A FLAMINGO object containing the resulted Interaction Frequency (IF) matrix, Pairwise Distance (PD) matrix and number of fragments.
}
\description{
Generate FLAMINGO input data from the .hic format data.
}
\examples{
flamingo.construct_flamingo_from_hic(hic_file="4DNFI8PZOJHN.hic",normalization='KR',resolution=5e3,chr_name='chr21',chr_size=46709983)
}
\keyword{FLAMINGO}
