% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo.impute_dist_from_epi.R
\name{flamingo.impute_dist_from_epi}
\alias{flamingo.impute_dist_from_epi}
\title{flamingo.impute_dist_from_epi}
\usage{
flamingo.impute_dist_from_epi(frag_res)
}
\arguments{
\item{frag_res}{Size of the small DNA fragment.}
}
\value{
Write out a list of files containing the genomic location and averaged epigenomic signals for DNA fragments.
}
\description{
Impute 3D distance between DNA fragments based on the 1D epigenomic signal
}
\examples{
flamingo.calculating_epigenomic_signal('DNase.bedgraph')
}
\keyword{FLAMINGO}
