% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_server.R
\docType{data}
\name{FLAMINGO_GUEST_ID}
\alias{FLAMINGO_GUEST_ID}
\title{FLAMINGO_GUEST_ID}
\format{An object of class \code{numeric} of length 1.}
\usage{
FLAMINGO_GUEST_ID
}
\value{
Set value.
}
\description{
Id to use for an unidentified user.
}
\keyword{datasets}
