% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoButton}
\alias{flamingoButton}
\title{Flamingo action button/link version}
\usage{
flamingoButton(inputId, label, icon = NULL, width = NULL,
  class = c("btn", "btn-primary"), ...)
}
\arguments{
\item{class}{HTML class attribute}

\item{...}{arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}}
}
\description{
Modified version of the default \code{\link[shiny:actionButton]{shiny::actionButton()}}
}
