% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{flamingoDB}
\alias{flamingoDB}
\title{Create a database settings object for the Flamingo Database}
\usage{
flamingoDB(server, port, database, uid, pwd, driver = "{FreeTDS}")
}
\arguments{
\item{server}{host name}

\item{port}{host port}

\item{database}{database name}

\item{uid}{username}

\item{pwd}{user pass}

\item{driver}{connection driver}
}
\value{
a database setttings object (list)
}
\description{
creates a database settings object which can then be used
to create new connections to the Flamingo Database
}
