% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{flamingoDBCheckPermissions}
\alias{flamingoDBCheckPermissions}
\title{flamingoDBCheckPermissions}
\usage{
flamingoDBCheckPermissions(dbSettings, user, resourceId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{user}{reactive expression yielding user}

\item{resourceId}{Resource identifier, e.g. \code{c("1000")}.}
}
\value{
Character vector of permission Modes, e.g. #' \code{c("CRUD" "R")}.
}
\description{
Consult the interface permissions for the user with given
user identifier by creating and executing a database query.
}
