% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{flamingoDBLogin}
\alias{flamingoDBLogin}
\title{Login against the database}
\usage{
flamingoDBLogin(dbSettings, uid, pwd)
}
\arguments{
\item{dbSettings}{object as returned by \code{\link[=flamingoDB]{flamingoDB()}}}

\item{uid}{user name}

\item{pwd}{user password}
}
\value{
user if succesful, -1 otherwise
}
\description{
Construct and execute a login query for the given credentials
}
