% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoHtmlDependencies}
\alias{flamingoHtmlDependencies}
\title{HTML Dependencies For The Flamingo Interface}
\usage{
flamingoHtmlDependencies()
}
\description{
HTML Dependencies For The Flamingo Interface
}
