% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoIncrementalPanel.R
\name{flamingoIncrementalPanel}
\alias{flamingoIncrementalPanel}
\title{flamingoIncrementalPanel}
\usage{
flamingoIncrementalPanel(input, output, session, panels_state,
  new_content_IDs, new_content_fun, ..., new_headings = NULL,
  collapsible = FALSE, show = TRUE)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{panels_state}{State of panel.}

\item{new_content_IDs}{ID with new content.}

\item{new_content_fun}{Function with new content.}

\item{new_headings}{New heading.}
}
\value{
smth.
}
\description{
flamingoIncrementalPanel
}
