% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoIncrementalPanel.R
\name{flamingoIncrementalPanelUI}
\alias{flamingoIncrementalPanelUI}
\title{flamingoIncrementalPanelUI}
\usage{
flamingoIncrementalPanelUI(id, ..., heading = NULL, footer = NULL,
  status = "default", collapsible = FALSE, show = TRUE,
  removable = TRUE)
}
\arguments{
\item{id}{module id}

\item{...}{Additional parameters.}

\item{heading}{NULL by default.}

\item{footer}{NULL by default.}

\item{status}{default.}

\item{collapsible}{FALSE.}

\item{show}{TRUE.}

\item{removable}{TRUE.}
}
\value{
List of tags.
}
\description{
flamingoIncrementalPanelUI
}
