% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoModule}
\alias{flamingoModule}
\title{Flamingo Module Template}
\usage{
flamingoModule(input, output, session, active, dbSettings, apiSettings,
  logMessage, logError)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{active}{reactive expression whether the module state should be updated.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{logMessage}{function that will be passed info messages}

\item{logError}{function that will be passed error messages}
}
\description{
Flamingo Module Template
}
