% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoModuleUI}
\alias{flamingoModuleUI}
\title{flamingoModuleUI}
\usage{
flamingoModuleUI(id)
}
\arguments{
\item{id}{Namespace ID for the module.}
}
\value{
No value.
}
\description{
Flamingo Module Template UI.
}
