% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{flamingoPanel}
\alias{flamingoPanel}
\title{flamingoPanel}
\usage{
flamingoPanel(id, ..., heading = NULL, footer = NULL,
  status = "default", collapsible = FALSE, show = TRUE)
}
\arguments{
\item{id}{Shiny module id.}

\item{...}{Additional parameters.}

\item{heading}{NULL by default.}

\item{footer}{NULL by default.}

\item{status}{default.}

\item{collapsible}{FALSE.}

\item{show}{TRUE.}
}
\description{
Adapted from shinyWidgets::panel() including IDs and collapsible
(inspired by shinyBS::bsCollapsePanel()).
}
