% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{flamingoPanelHeading}
\alias{flamingoPanelHeading}
\title{flamingoPanelHeading}
\usage{
flamingoPanelHeading(heading)
}
\arguments{
\item{heading}{title of the panel}
}
\value{
heading
}
\description{
flamingoPanelHeading
}
