% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{flamingoServer}
\alias{flamingoServer}
\title{flamingoServer}
\usage{
flamingoServer(host, port, scheme = c("http", "https"))
}
\arguments{
\item{host}{Host name.}

\item{port}{Host port.}

\item{scheme}{Communication scheme.}
}
\value{
A flamingo server settings object (HTTP by default).
}
\description{
Creates a settings object which can then be used to connect
to the Flamingo API Server.
}
