% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoTable.R
\name{flamingoTable}
\alias{flamingoTable}
\title{Module for flamingoTable}
\usage{
flamingoTable(input, output, session, data, selection = "none",
  escape = TRUE, scrollX = FALSE, filter = FALSE, rownames = FALSE,
  colnames = TRUE, preselRow = reactive({     NULL }),
  maxrowsperpage = 10, logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{logMessage}{function that will be passed info messages}
}
\value{
null
}
\description{
Server logic to show the flamingo table output
}
