% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoTableButton}
\alias{flamingoTableButton}
\title{flamingoTableButton}
\usage{
flamingoTableButton(inputId, label, icon = NULL, width = NULL,
  class = c("btn", "btn-primary"), ...)
}
\arguments{
\item{...}{Arguments to \code{\link[flamingo:flamingoButton]{flamingo::flamingoButton()}}.}
}
\value{
Acces to Oasis UI.
}
\description{
Modified version of the default \code{\link[flamingo:flamingoButton]{flamingo::flamingoButton()}}.
}
