% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexColumn.R
\name{flexColumn}
\alias{flexColumn}
\alias{flexColumnUI}
\title{Resizable UI column}
\usage{
flexColumn(input, output, session, width)

flexColumnUI(id, width, ...)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{width}{Width of the column, a reactive value for \code{flexColumn()},
non-reactive (or \link[shiny:isolate]{isolated}) for \code{flexColumnUI()}}

\item{id}{Namespace ID for the module.}
}
\description{
Shiny module providing support for dynamic UI \link[shiny:column]{column} width
without re-rendering the column content.
}
\details{
Dynamic column width via \code{renderUI()} based on a reactive value forces the
re-rendering of the column content. This has the following undesired effects:
\itemize{
\item The state of the existing column UI content (including user input) is
lost.
\item Depending on the complexity of the content, rendering can slow-down the
re-sizing.
}

The \emph{flexColumn} Shiny module prevents re-rendering based on reactive column
width. Instead, it changes the class of the existing column UI element
(\code{"col-sm-<WIDTH>"}) depending on a reactive \code{width} value.

The module relies on \code{\link[shinyjs:addClass]{shinyjs::addClass()}} and \code{\link[shinyjs:removeClass]{shinyjs::removeClass()}},
which means \code{\link[shinyjs:useShinyjs]{shinyjs::useShinyjs()}} must be included in the app \code{ui}
definition.
}
\examples{
if (interactive()) {
  library(shiny)
  twoFlexColumnsUI <- function(sidebar_w) {
    # make the UI rendering slow
    Sys.sleep(1)
    fluidRow(
      flexColumnUI(
        id = "sidebar", width = sidebar_w,
        radioButtons("radio", "Does radio selection persist re-sizing?", LETTERS[1:3])
      ),
      flexColumnUI(
        id = "main", width = 12L - sidebar_w,
        textInput("text", "Does text input persist re-sizing?", width = "100\%")
      )
    )
  }
  ui <- fluidPage(
    shinyjs::useShinyjs(),
    titlePanel("Resizable columns with `flexColumn` module"),
    fluidPage(
      fluidRow(sliderInput("sidebar_w", "Sidebar Width", 1L, 11L, 3L, 1L)),
      fluidRow(checkboxInput("flex", "Use flexColumn", FALSE)),
      fluidRow(uiOutput("twocols"))
    )
  )
  server <- function(input, output) {
    observe(
      if (input$flex) {
        # UI rendering non-reactive to input$sidebar_w
        output$twocols <-
          renderUI(twoFlexColumnsUI(isolate(input$sidebar_w)))
        # reactivity via flexColumn module
        callModule(flexColumn, "sidebar", reactive(input$sidebar_w))
        callModule(flexColumn, "main", reactive(12L - input$sidebar_w))
      } else {
        # UI rendering reactive to input$sidebar_w
        output$twocols <-
          renderUI(twoFlexColumnsUI(input$sidebar_w))
      }
    )
  }
  shinyApp(ui, server)
}
}
