% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatedinputs_module.R
\name{generatedinputsUI}
\alias{generatedinputsUI}
\alias{generatedinputs}
\title{generatedinputsUI}
\usage{
generatedinputsUI(id)

generatedinputs(input, output, session, analysisID, counter = NULL,
  active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis id.}

\item{counter}{Reactive value storing actionButton status.}
}
\value{
List of tags.
}
\description{
UI/View for uploaded inputs of an analysis.

Server logic for uploaded inputs of an analysis.
}
