% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{get_analyses_inputs_tar}
\alias{get_analyses_inputs_tar}
\title{Get analysis inputs tar path}
\usage{
get_analyses_inputs_tar(id, destdir = tempdir())
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{destdir}{path where to download tar file.}
}
\value{
Path to file downloaded.
}
\description{
Downloads the analysis inputs tar and returns download path.
}
