% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{get_analyses_tar}
\alias{get_analyses_tar}
\title{Get analysis tar path}
\usage{
get_analyses_tar(id, label, destdir = tempdir(), oasisapi)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{label}{input/output}

\item{destdir}{path where to download tar file.}

\item{oasisapi}{as stored in session$userData$oasisapi.}
}
\value{
Path to file downloaded.
}
\description{
Downloads the analysis tar and returns download path.
}
