% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getAccountName}
\alias{getAccountName}
\title{getAccountName}
\usage{
getAccountName(dbSettings)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}
}
\value{
Account name; \code{data.frame} of 2 variables:
\itemize{
\item \code{Account ID}
\item \code{progOasisId}
}.
}
\description{
Get Account Names.
}
