% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getDeptData}
\alias{getDeptData}
\title{getDeptData}
\usage{
getDeptData(dbSettings, user)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{user}{reactive expression yielding user}
}
\description{
Queries the database to retrieve department login and password
details for the designated user.
}
