% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getEventOccurrence}
\alias{getEventOccurrence}
\title{getEventOccurrence}
\usage{
getEventOccurrence(dbSettings, prgoasisid, simplify = TRUE)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{prgoasisid}{Oasis programme id.}

\item{simplify}{logical}
}
\value{
Event occurrence.
}
\description{
Get the event occurrence list.
}
