% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getEventSet}
\alias{getEventSet}
\title{getEventSet}
\usage{
getEventSet(dbSettings, prgoasisid, simplify = TRUE)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{prgoasisid}{Oasis programme id.}

\item{simplify}{logical}
}
\value{
Event set.
}
\description{
Get the event set list.
}
