% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileList}
\alias{getFileList}
\title{Get Output Files}
\usage{
getFileList(dbSettings, prrunid)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{prrunid}{Process run id.}
}
\value{
Output files; \code{data.frame} of 10 variables:
\itemize{
\item \code{FileID}
\item \code{File Name}
\item \code{Description}
\item \code{Location}
\item \code{Location Unix}
\item \code{File Type}
\item \code{Owner}
\item \code{Source}
\item \code{Resource Table}
\item \code{Resource Key}
}.
}
\description{
Get the list of output files for a given process run id.
}
