% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileLocationPath}
\alias{getFileLocationPath}
\title{getFileLocationPath}
\usage{
getFileLocationPath(dbSettings, fileType)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{fileType}{File type.}
}
\value{
Path.
}
\description{
Get location path for a given fileType.
}
