% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileSourceLocationFile}
\alias{getFileSourceLocationFile}
\title{getFileSourceLocationFile}
\usage{
getFileSourceLocationFile(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
file source\code{data.frame} of 2 variables:
\itemize{
\item \code{FileName}
\item \code{FileId}
}.
}
\description{
Get File Source Location File.
}
