% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileSourceReinsuranceScopeFile}
\alias{getFileSourceReinsuranceScopeFile}
\title{getFileSourceReinsuranceScopeFile}
\usage{
getFileSourceReinsuranceScopeFile(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
\code{data.frame} of 2 variables:
\itemize{
             \item \code{FileName}
             \item \code{FileId}
}.
}
\description{
Get File Source Reinsurance Scope File.
}
