% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{getGeoDetails}
\alias{getGeoDetails}
\title{getGeoDetails}
\usage{
getGeoDetails(address)
}
\arguments{
\item{address}{Server address.}
}
\value{
\code{data.frame} with details. If no match is found, the result
is a \code{data.frame} of \code{NA} values.
}
\description{
This function can be used to query google servers for
geo details.
}
