% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelAnaDetailsForFileOutput}
\alias{getModelAnaDetailsForFileOutput}
\title{getModelAnaDetailsForFileOutput}
\usage{
getModelAnaDetailsForFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{processRunId}{identifier selected model analysis.}
}
\value{
Model analysis details.
}
\description{
Get Model Analysis Details.
}
