% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelAnaParamFileOutput}
\alias{getModelAnaParamFileOutput}
\title{getModelAnaParamFileOutput}
\usage{
getModelAnaParamFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{processRunId}{Reactive string expression for reselected run id from defineProgramme.}
}
\value{
Param details.
}
\description{
Get Model Runtime Param Details.
}
