% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelData}
\alias{getModelData}
\title{getModelData}
\usage{
getModelData(dbSettings, pruser, prmodel, prprogramme, prworkflow)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{pruser}{Model user.}

\item{prmodel}{Model model.}

\item{prprogramme}{Model programme.}

\item{prworkflow}{Model workflow.}
}
\value{
Model data; \code{data.frame} of 3 variables:
\itemize{
\item \code{ProgOasisId}
\item \code{ProgName}
\item \code{ModelName}
}.
}
\description{
Get Model Data.
}
