% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelList}
\alias{getModelList}
\title{getModelList}
\usage{
getModelList(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
Res \code{data.frame} of 3 variables:
\itemize{
\item \code{Model ID}
\item \code{Model Name}
\item \code{Model Description}
}.
}
\description{
Get Model Data.
}
