% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelRun}
\alias{getModelRun}
\title{getModelRun}
\usage{
getModelRun(dbSettings, modelid, prstatus)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{modelid}{Model id.}

\item{prstatus}{Model status.}
}
\description{
Get Model Run.
}
