% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getModelRunDetails}
\alias{getModelRunDetails}
\title{getModelRunDetails}
\usage{
getModelRunDetails(dbSettings, wfid)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{wfid}{Workflow id.}
}
\value{
Log details; \code{data.frame} of 5 variables:
\itemize{
\item \code{Element ID}
\item \code{Element Name}
\item \code{Status}
\item \code{Description}
\item \code{CompletedAt}
}.
}
\description{
Get log details for a run.
}
