% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getOasisSystemId}
\alias{getOasisSystemId}
\title{getOasisSystemId}
\usage{
getOasisSystemId(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
Oasis systems; \code{data.frame} of 2 variables:
\itemize{
\item \code{OasisSystemID}
\item \code{OasisSystemName}
}.
}
\description{
Get Oasis Systems.
}
