% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getOutputOptions}
\alias{getOutputOptions}
\title{Get the list of output presets}
\usage{
getOutputOptions(dbSettings, simplify = TRUE)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{simplify}{if \code{TRUE}, a vector will be returned instead of
a \code{data.frame} only a single column is returned.}
}
\value{
output options
}
\description{
Get the list of output presets
}
