% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcessData}
\alias{getProcessData}
\title{getProcessData}
\usage{
getProcessData(dbSettings, pruser, prmodel, prprogramme, prworkflow)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{pruser}{Process user.}

\item{prmodel}{Process model.}

\item{prprogramme}{Process programme.}

\item{prworkflow}{Process workflow.}
}
\value{
Process data; \code{data.frame} of 3 variables:
\itemize{
\item \code{ProgOasisId}
\item \code{ProgName}
\item \code{ModelName}
}.
}
\description{
Get Process Data.
}
