% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcessRun}
\alias{getProcessRun}
\title{Get Process Run}
\usage{
getProcessRun(dbSettings, procid, prstatus)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{procid}{process id}

\item{prstatus}{process status}
}
\description{
Get Process Run
}
