% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunDetForFileOutput}
\alias{getProcRunDetForFileOutput}
\title{getProcRunDetForFileOutput}
\usage{
getProcRunDetForFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{processRunId}{identifier selected model analysis.}
}
\value{
Model run details.
}
\description{
Get Model Run Details.
}
