% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunParamFileOutput}
\alias{getProcRunParamFileOutput}
\title{getProcRunParamFileOutput}
\usage{
getProcRunParamFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{processRunId}{Reactive string expression for reselected run id from defineProgramme.}
}
\value{
Param details.
}
\description{
Get Process Runtime Param Details.
}
