% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProgOasisForProgdata}
\alias{getProgOasisForProgdata}
\title{Get Programme Oasis for a given Programme Id}
\usage{
getProgOasisForProgdata(dbSettings, progId)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}
}
\value{
\code{data.frame} of 11 variables:
\itemize{
		\item \code{ProgOasisId}
		\item \code{ProgName}
		\item \code{ModelName}
		\item \code{TransformName} 
		\item \code{SourceFileId}
		\item \code{FileID} 
		\item \code{Status}
		\item \code{API1aDateTime} 
		\item \code{API1bDateTime}
		\item \code{API1cDateTime} 
		\item \code{SessionId}
}
}
\description{
Run getProgOasisForProg against the database.
}
