% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProgrammeList}
\alias{getProgrammeList}
\title{getProgrammeList}
\usage{
getProgrammeList(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
\code{data.frame} of 7 variables:
\itemize{
\item \code{Programme ID}
\item \code{Programme Name}
\item \code{Account ID}
\item \code{Account Name}
\item \code{Transform ID}
\item \code{Transform}
\item \code{Status}
}.
}
\description{
Get a list of programme data.
}
