% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getResourceType}
\alias{getResourceType}
\title{getResourceType}
\usage{
getResourceType(dbSettings)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}
}
\value{
Resource types; \code{data.frame} of 3 variables:
\itemize{
\item \code{ResourceTypeID}
\item \code{ResourceTypeName}
\item \code{Source}
}.
}
\description{
Fetch resource type name and ID.
}
