% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_utils.R
\name{getTableOptions}
\alias{getTableOptions}
\title{Get table options.}
\usage{
getTableOptions(scrollX, maxrowsperpage, filter, escape)
}
\arguments{
\item{scrollX}{Param to allow scrollX.}

\item{maxrowsperpage}{Maximum number of rows to display per page.}

\item{filter}{Show or hide filter.}

\item{escape}{Param to avoid escape row.}
}
\description{
Returns option list for datatable.
}
