% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getTransformNameCanModel}
\alias{getTransformNameCanModel}
\title{getTransformNameCanModel}
\usage{
getTransformNameCanModel(dbSettings)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
Transforms; \code{data.frame} of 2 variables:
\itemize{
\item \code{TransformName}
\item \code{TransformID}
}.
}
\description{
Get Transforms from Canonical to Model.
}
