% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getTransformNameSourceCan}
\alias{getTransformNameSourceCan}
\title{getTransformNameSourceCan}
\usage{
getTransformNameSourceCan(dbSettings)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}
}
\value{
Transforms; \code{data.frame} of 2 variables:
\itemize{
\item \code{TransformName}
\item \code{TransformID}
}.
}
\description{
Get Transforms from Source to Canonical.
}
