% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{loadProgrammeData}
\alias{loadProgrammeData}
\title{loadProgrammeData}
\usage{
loadProgrammeData(apiSettings, progId, logWarning = function(w) warning,
  ...)
}
\arguments{
\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{progId}{programme identifier-}

\item{logWarning}{Warning message callback-}

\item{...}{Additional arguments to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
http status category warn_for_status.
}
\description{
A function to load Programme data.
}
