% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{loadProgrammeModel}
\alias{loadProgrammeModel}
\title{loadProgrammeModel}
\usage{
loadProgrammeModel(apiSettings, progOasisId, logWarning = function(w)
  warning, ...)
}
\arguments{
\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{progOasisId}{Oasis programme identifier-}

\item{logWarning}{Warning message callback-}

\item{...}{Additional arguments to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
http status category warn_for_status.
}
\description{
A function to load Programme data.
}
