% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_server.R
\name{modelSupplierPage}
\alias{modelSupplierPage}
\title{modelSupplierPage}
\usage{
modelSupplierPage(input, output, session, dbSettings,
  logMessage = message, active = reactive(TRUE))
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{logMessage}{Function that will be passed info messages.}

\item{active}{Reactive expression whether the module state should be updated.}
}
\value{
Empty list.
}
\description{
Server logic for the model supplier page.
}
