% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo-package.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate in operator}
\usage{
... \%notin\% NA
}
\arguments{
\item{...}{extra arguments to \code{\link[=Negate]{Negate()}}}

\item{NA}{reqested by check_man}
}
\description{
negate in operator
}
