% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfiles_module.R
\name{outputfilesUI}
\alias{outputfilesUI}
\alias{outputfiles}
\title{outputfilesUI}
\usage{
outputfilesUI(id)

outputfiles(input, output, session,
  tbl_filesListDataana = reactive(NULL),
  tbl_filesListDatapf = reactive(NULL), anaId = reactive(""),
  portfolioId = reactive(""), active, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{tbl_filesListDataana}{tbl of output files to view}

\item{tbl_filesListDatapf}{dataframe of input files to view}

\item{anaId}{id of analysis}

\item{portfolioId}{id of portfolio associated to the analysis}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.
}
\description{
UI/View for output files of an analysis.

Server logic for output files of an analysis.
}
