% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfiles_module.R
\name{outputfilesUI}
\alias{outputfilesUI}
\title{outputfiles_ui
Run outputfiles UI}
\usage{
outputfilesUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
output files of a Run
}
