% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveNavigation.R
\name{outputNavigation}
\alias{outputNavigation}
\title{outputNavigation}
\usage{
outputNavigation(state, force_react = TRUE)
}
\arguments{
\item{state}{Reactive navigation state created with \code{reactiveNavigation()}.}

\item{force_react}{Logical flag specifying whether the observed event for
navigation state update should always invalidate the reactive navigation
state, thus forcing rectivity. This is especially desirable within inner
modules.}
}
\value{
\code{outputNavigation()} returns a list containing the reactive
navigation state, to be possibliy \code{c}ombined with other module outputs. The
reactive navigation state can then be extracted using \code{getNavigation()}.
}
\description{
outputNavigation
}
