% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagestructureSidebar.R
\name{pagestructureSidebar}
\alias{pagestructureSidebar}
\title{pagestructureSidebar}
\usage{
pagestructureSidebar(ns = identity, collapsed = FALSE)
}
\arguments{
\item{ns}{ns.}

\item{collapsed}{FALSE.}
}
\value{
Side bar in pagestructure.
}
\description{
pagestructureSidebar
}
