% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panel_configureAdvancedGUL}
\alias{panel_configureAdvancedGUL}
\title{panel_configureAdvancedGUL}
\usage{
panel_configureAdvancedGUL(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration GUL.
}
