% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panel_configureAdvancedIL}
\alias{panel_configureAdvancedIL}
\title{panel_configureAdvancedIL}
\usage{
panel_configureAdvancedIL(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration IL.
}
