% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panel_configureAdvancedRI}
\alias{panel_configureAdvancedRI}
\title{panel_configureAdvancedRI}
\usage{
panel_configureAdvancedRI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration RI.
}
