% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseAnalysis_module.R
\name{panelModelTable}
\alias{panelModelTable}
\title{panelAnalysisGenInputs}
\usage{
panelModelTable(id)
}
\arguments{
\item{id}{module id}

\item{id}{module id}
}
\description{
Function wrapping panel to show analyses generated inputs table.

Function wrapping panel to show list of models table.
}
