% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseAnalysis_module.R
\name{panelAnalysisLog}
\alias{panelAnalysisLog}
\title{panelAnalysisLog}
\usage{
panelAnalysisLog(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show analyses logs table.
}
