% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panelAnalysisLogs}
\alias{panelAnalysisLogs}
\title{panelAnalysisLogs}
\usage{
panelAnalysisLogs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show log table for specific Process Run.
}
