% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel.R
\name{panelAssociateModel}
\alias{panelAssociateModel}
\title{Function wrapping panel to associate model}
\usage{
panelAssociateModel(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to associate model
}
