% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelDefineIDs}
\alias{panelDefineIDs}
\title{Function wrapping panel to define prgramme and model IDs}
\usage{
panelDefineIDs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to define prgramme and model IDs
}
