% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelDefineOutputConfiguration}
\alias{panelDefineOutputConfiguration}
\title{Function wrapping sub-panel to define outputs configuration}
\usage{
panelDefineOutputConfiguration(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs configuration
}
