% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelDefineOutputs}
\alias{panelDefineOutputs}
\title{Function wrapping panel to define outputs}
\usage{
panelDefineOutputs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to define outputs
}
