% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{panelDefineOutputsDetails}
\alias{panelDefineOutputsDetails}
\title{Function wrapping sub-panel to define outputs details}
\usage{
panelDefineOutputsDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs details
}
