% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelDefineProgramme}
\alias{panelDefineProgramme}
\title{Function wrapping panel to create/amend programme}
\usage{
panelDefineProgramme(id)
}
\description{
Function wrapping panel to create/amend programme
}
