% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel_module.R
\name{panelModelTable}
\alias{panelModelTable}
\title{panelModelTable}
\usage{
panelModelTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show list of models table.
}
