% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_ui.R
\name{panelOutputModuleUI}
\alias{panelOutputModuleUI}
\title{Function wrapping panel to define prgramme and model IDs}
\usage{
panelOutputModuleUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to define prgramme and model IDs
}
