% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_choosePortfolio_module.R
\name{panelPortfolioDetails}
\alias{panelPortfolioDetails}
\title{panelPortfolioDetails}
\usage{
panelPortfolioDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show details of portfolio.
}
