% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_choosePortfolio_module.R
\name{panelPortfolioTable}
\alias{panelPortfolioTable}
\title{panelPortfolioTable}
\usage{
panelPortfolioTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show created portfolios table.
}
