% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelProcessRunLogs}
\alias{panelProcessRunLogs}
\title{Function wrapping panel to show Log Process Run table}
\usage{
panelProcessRunLogs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show Log Process Run table
}
