% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelProcessRunTable}
\alias{panelProcessRunTable}
\title{Function wrapping panel to show process run table}
\usage{
panelProcessRunTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show process run table
}
