% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{panelProgrammeDetails}
\alias{panelProgrammeDetails}
\title{panelProgrammeDetails}
\usage{
panelProgrammeDetails(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\description{
Function wrapping panel to show details of programme.
}
