% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel_module.R
\name{panelProgrammeModelTable}
\alias{panelProgrammeModelTable}
\title{panelProgrammeModelTable}
\usage{
panelProgrammeModelTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show created programme model table.
}
