% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R
\name{panelSummaryTableModule}
\alias{panelSummaryTableModule}
\title{Module for Summary Table Panel}
\usage{
panelSummaryTableModule(input, output, session, dbSettings, apiSettings,
  userId, logMessage = message, selectRunID)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{logMessage}{Function that will be passed info messages.}

\item{selectRunID}{selected runID}
}
\value{
null
}
\description{
Server logic to show the summary table output
}
