% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_server.R
\name{panelViewOutputFilesModule}
\alias{panelViewOutputFilesModule}
\title{Module for the Panel to View Output Files}
\usage{
panelViewOutputFilesModule(input, output, session, logMessage = message,
  filesListData)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{logMessage}{function that will be passed info messages}

\item{filesListData}{table of output files for a given runID}
}
\value{
list of reactives:
}
\description{
Server logic to view output files
}
