% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGUL}
\alias{plotGUL}
\title{GUL plot}
\usage{
plotGUL(outputPlotData, interactive = FALSE, ...)
}
\arguments{
\item{outputPlotData}{data for plot}

\item{interactive}{create interactive plot using \link{plotly::plot_ly}}

\item{...}{extra arguments to \link{plot}}
}
\value{
nothing; the interactive plot object if \code{interactive = TRUE}
}
\description{
plot GUL using the current device
}
