% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadedinputs_module.R
\name{portfolio_detailsUI}
\alias{portfolio_detailsUI}
\alias{portfolio_details}
\title{portfolio_detailsUI}
\usage{
portfolio_detailsUI(id)

portfolio_details(input, output, session, portfolioID, counter = NULL,
  active = reactive(TRUE))
}
\arguments{
\item{portfolioID}{selected portfolio ID.}
}
\value{
List of tags.
}
\description{
UI/View for portfolio details.

Server logic for portfolio details.
}
