% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRun_server.R, R/processRun_ui.R
\name{processRunPage}
\alias{processRunPage}
\alias{processRunPageUI}
\title{Process Run Page}
\usage{
processRunPage(input, output, session, dbSettings, apiSettings, userId,
  preselRunId = reactive(-1), preselProcId = reactive(-1),
  progOasisId = reactive(-1), active = reactive(TRUE),
  logMessage = message, reloadMillis = 10000)

processRunPageUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{reactive expresssion yielding user id}

\item{preselRunId}{reactive expresssion yielding preselected run id}

\item{preselProcId}{reactive expression yielding preselected proc id}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{module id}
}
\value{
empty list

list of tags
}
\description{
Server logic for the process run page

UI/View for the process run page
}
