% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRun_ui.R
\name{processRunParamAdvancedILUI}
\alias{processRunParamAdvancedILUI}
\title{UI/View to define the Advanced  IL parameters of a process run}
\usage{
processRunParamAdvancedILUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  IL parameters of a process run
}
