% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRun_ui.R
\name{processRunParamPerilAdvancedUI}
\alias{processRunParamPerilAdvancedUI}
\title{UI/View to define the Advanced  Model perils of a process run}
\usage{
processRunParamPerilAdvancedUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI/View to define the Advanced  Model perils of a process run
}
