% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionBatch_server.R,
%   R/programmeDefinitionBatch_ui.R
\name{programmeDefinitionBatch}
\alias{programmeDefinitionBatch}
\alias{programmeDefinitionBatchUI}
\title{programmeDefinitionBatch_server}
\usage{
programmeDefinitionBatch(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

programmeDefinitionBatchUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{id}{Shiny module id.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{programmeDefinitionBatch()}, list of reactives.

List of tags.
}
\description{
Server logic to define a programme.

UI/View for the process run page.
}
