% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_server.R,
%   R/programmeDefinitionSingle_ui.R
\name{programmeDefinitionSingle}
\alias{programmeDefinitionSingle}
\alias{programmeDefinitionSingleUI}
\title{programmeDefinitionSingle}
\usage{
programmeDefinitionSingle(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  preselRunId = reactive(-1), preselProcId = reactive(-1),
  preselPanel = reactive(1), reloadMillis = 10000)

programmeDefinitionSingleUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{programmeDefinitionSingle()}, list of reactives.

List of tags.
}
\description{
Server logic to define a programme.

UI/View for the process run page
}
