% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleProgrammeWorkflowSteps.R
\docType{data}
\name{programmeWorkflowSteps}
\alias{programmeWorkflowSteps}
\title{programmeWorkflowSteps}
\format{An object of class \code{list} of length 3.}
\usage{
programmeWorkflowSteps
}
\value{
List.
}
\description{
programmeWorkflowSteps
}
\keyword{datasets}
