% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveConditionalPanels.R
\name{reactiveConditionalPanelsUI}
\alias{reactiveConditionalPanelsUI}
\title{reactiveConditionalPanelsUI}
\usage{
reactiveConditionalPanelsUI(id, panels)
}
\arguments{
\item{panels}{Named list of UI content for a set of panels. Only the panel
whose name matches the \code{visible} reactive input to the module server
function is shown.}
}
\value{
Panels.
}
\description{
reactiveConditionalPanelsUI
}
