% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{renderflamingoPanelHeading}
\alias{renderflamingoPanelHeading}
\title{renderflamingoPanelHeading}
\usage{
renderflamingoPanelHeading(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{expression}

\item{env}{environment}

\item{...}{additional parameters}
}
\description{
renderflamingoPanelHeading
}
