% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{return_accounts_file_df}
\alias{return_accounts_file_df}
\title{Return accounts_file Dataframe}
\usage{
return_accounts_file_df(id)
}
\arguments{
\item{id}{A unique integer value identifying the portfolio.}
}
\value{
Dataframe of accounts_file.
}
\description{
Returns a dataframe of accounts_file.
}
