% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{return_analyses_df}
\alias{return_analyses_df}
\title{Return analyses Dataframe}
\usage{
return_analyses_df(name = "")
}
\arguments{
\item{name}{name of the analysis.}
}
\value{
dataframe of previously posted analyses Default empty string returns all analyses.
}
\description{
Returns a dataframe of analyses
}
