% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{return_analyses_id_df}
\alias{return_analyses_id_df}
\title{Return analysis details Dataframe}
\usage{
return_analyses_id_df(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
dataframe of details of previously posted analysis
}
\description{
Returns a dataframe of analysis details
}
