% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_input_file_df}
\alias{return_analyses_input_file_df}
\title{Return analyses input files  Dataframe}
\usage{
return_analyses_input_file_df(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
dataframe of input files.
}
\description{
Returns a dataframe of input files
}
