% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_input_file_wicons_df}
\alias{return_analyses_input_file_wicons_df}
\title{Return analyses input files dataframe with icons}
\usage{
return_analyses_input_file_wicons_df(id, data_hub, oasisapi)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{oasisapi}{as stored in session$userData$oasisapi.}

\item{datahub}{as stored in session$userData$datahub}
}
\value{
Dataframe of input files with icons.
}
\description{
Returns a dataframe of input files with icons.
}
