% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_output_file_df}
\alias{return_analyses_output_file_df}
\title{Return analyses output files  Dataframe}
\usage{
return_analyses_output_file_df(id)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}
}
\value{
Dataframe of output files.
}
\description{
Returns a dataframe of output files.
}
