% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_run_df}
\alias{return_analyses_run_df}
\title{Return analyses run Dataframe}
\usage{
return_analyses_run_df(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
dataframe of previously posted analysis states.
}
\description{
Returns a dataframe of analyses after run started
}
