% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_spec_input_file_df}
\alias{return_analyses_spec_input_file_df}
\title{Return specific analyses input file as Dataframe}
\usage{
return_analyses_spec_input_file_df(id, fileName)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{fileName}{name of file to read}
}
\value{
dataframe of specific input file.
}
\description{
Returns a dataframe of specific input file
}
