% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_analyses_spec_output_file_df}
\alias{return_analyses_spec_output_file_df}
\title{Return specific analyses output file as dataframe}
\usage{
return_analyses_spec_output_file_df(id, fileName)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{fileName}{Name of file to read.}
}
\value{
Dataframe of specific output file.
}
\description{
Returns a dataframe of a specific output file.
}
