% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{return_df}
\alias{return_df}
\title{Return Dataframe from API response}
\usage{
return_df(api_query, api_param = "")
}
\arguments{
\item{api_query}{function representing the API query}

\item{api_param}{parameter for the api_query}
}
\value{
dataframe of the API response
}
\description{
Returns a dataframe of the API response
}
