% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{return_file_df}
\alias{return_file_df}
\title{Return file as Dataframe}
\usage{
return_file_df(api_query, api_param = "")
}
\arguments{
\item{api_query}{Function representing the API query.}

\item{api_param}{Parameter for the api_query.}
}
\value{
Dataframe of a file downloaded from the API.
}
\description{
Returns a dataframe of a file downloaded from the API.
}
