% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{return_input_generation_traceback_file_df}
\alias{return_input_generation_traceback_file_df}
\title{Return input_generation_traceback_file Dataframe}
\usage{
return_input_generation_traceback_file_df(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
dataframe of iinput_generation_traceback_file
}
\description{
Returns a dataframe of input_generation_traceback_file
}
