% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{return_location_file_df}
\alias{return_location_file_df}
\title{Return location_file Dataframe}
\usage{
return_location_file_df(id)
}
\arguments{
\item{id}{A unique integer value identifying the portfolio.}
}
\value{
Dataframe of location_file.
}
\description{
Returns a dataframe of location_file
}
