% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{return_model_df}
\alias{return_model_df}
\title{Return Model Dataframe}
\usage{
return_model_df(id)
}
\arguments{
\item{id}{the ID for the model.}
}
\value{
dataframe of previously posted model.
}
\description{
Returns a dataframe of model
}
