% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{return_models_df}
\alias{return_models_df}
\title{Return Models Dataframe}
\usage{
return_models_df(supplier_id = "")
}
\arguments{
\item{supplier_id}{the supplier ID for the model. Default is empty string.}
}
\value{
dataframe of previously posted models. Default empty string returns all portfolios.
}
\description{
Returns a dataframe of models
}
