% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{return_portfolio_details_df}
\alias{return_portfolio_details_df}
\title{Return Details Portfolio id Dataframe}
\usage{
return_portfolio_details_df(id)
}
\arguments{
\item{id}{id of the portfolio.}
}
\value{
dataframe of details of previously posted portfolio
}
\description{
Returns a dataframe of portfolio's details
}
