% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{return_portfolios_df}
\alias{return_portfolios_df}
\title{Return Portfolios Dataframe}
\usage{
return_portfolios_df(name = "")
}
\arguments{
\item{name}{name of the portfolio.}
}
\value{
dataframe of previously posted portfolios. Default empty string returns all portfolios.
}
\description{
Returns a dataframe of portfolios
}
