% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{return_portfolios_stored_name}
\alias{return_portfolios_stored_name}
\title{Return portfolio file stored name}
\usage{
return_portfolios_stored_name(id, filetype)
}
\arguments{
\item{id}{A unique integer value identifying this portfolio.}

\item{filetype}{character string indicating the type of file: can be location_file, accounts_file, reinsurance_info_file or reinsurance_scope_file.}
}
\value{
The stored name of the file.
}
\description{
Returns file stored name
}
