% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{return_response}
\alias{return_response}
\title{Return response from API query}
\usage{
return_response(api_query, api_param = "")
}
\arguments{
\item{api_query}{Function representing the API query.}

\item{api_param}{Parameter for the api_query.}
}
\value{
List of the API response.
}
\description{
Returns a list of the API response.
}
