% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{return_tbl_analysesData}
\alias{return_tbl_analysesData}
\title{Return analyses data for DT}
\usage{
return_tbl_analysesData(name = "", oasisapi)
}
\arguments{
\item{name}{Name of the analyses.}

\item{oasisapi}{as stored in session$userData$oasisapi}
}
\value{
Dataframe of previously posted analyses. Default empty string returns all analyses.
}
\description{
Returns a dataframe of analyses ready for being rendered as a data table.
}
