% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{return_tbl_analysesData_nice}
\alias{return_tbl_analysesData_nice}
\title{Return analysis table in nice format for DT}
\usage{
return_tbl_analysesData_nice(tbl_analysesData, oasisapi)
}
\arguments{
\item{tbl_analysesData}{dataframe of analyses}

\item{oasisapi}{as stored in session$userData$oasisapi}
}
\value{
Dataframe of analyses
}
\description{
Returns a dataframe of analyses prettified for being rendered as a data table.
}
