% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{return_tbl_modelsData}
\alias{return_tbl_modelsData}
\title{Return Models Data fot DT}
\usage{
return_tbl_modelsData(supplier_id = "")
}
\arguments{
\item{supplier_id}{the supplier ID for the model. Default is empty string.}
}
\value{
dataframe of previously posted modelss. Default empty string returns all models.
}
\description{
Returns a dataframe of models ready for being rendered as a data table
}
