% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolio.R
\name{return_tbl_portfolioDetails}
\alias{return_tbl_portfolioDetails}
\title{Return Portfolio Details fot DT}
\usage{
return_tbl_portfolioDetails(id)
}
\arguments{
\item{id}{id of the portfolio.}
}
\value{
dataframe of details of previously posted portfolio.
}
\description{
Returns a dataframe of portfolio's details ready for being rendered as a data table
}
