% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{run_variable}
\alias{run_variable}
\title{Generate Random Distribution}
\usage{
run_variable(n, mean, sd, method = "RND", lhs = NULL, distro = "normal",
  lower = -Inf, upper = Inf, seed = 1234)
}
\arguments{
\item{n}{number of realizations}

\item{mean}{mean}

\item{sd}{standard deviation}

\item{method}{sampling method, either "RND" or "LHS". Default = "RND"}

\item{lhs}{vector of quatiles (obtained from LH sampling functions). Not required
is \code{method = "RND"}. Default = NULL}

\item{distro}{distribution}

\item{lower}{lower limit of the truncated distribution}

\item{upper}{upper limit of the truncated distribution}

\item{seed}{integer serving as seed number}
}
\value{
an array with n random values for the given parameters
}
\description{
Generate Random Distribution
}
\examples{
\dontrun{
distribution_1 <- run_variable(100, 10, 4)
hist(distribution_1)

distribution_2 <- run_variable(100, 10, 4, "truncated normal", 7, 13, seed = 123)
hist(distribution_2)
}

}
