% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(apiSettings, anaid, logWarning = warning, ...)
}
\arguments{
\item{apiSettings}{as returned from \link{flamingoServer}}

\item{anaid}{Analysis identifier.}

\item{logWarning}{Warning message callback-}

\item{...}{Further arguments to \link{GET}.}
}
\value{
http status category warn_for_status.
}
\description{
Submit a GET request to the Flamingo API Server to execute
the workflow.
}
