% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{runProcess}
\alias{runProcess}
\title{Call BFE webservice to initiate a job}
\usage{
runProcess(apiSettings, runId, logWarning = warning, ...)
}
\arguments{
\item{apiSettings}{settings object as returned by \link{flamingoServer}}

\item{runId}{run identifier}

\item{logWarning}{warning message callback}

\item{...}{further arguments to \link{GET}}
}
\value{
http status category warn_for_status
}
\description{
Submit a GET request to the Flamingo API Server to execute 
the workflow.
}
