% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{set_extractFilePath}
\alias{set_extractFilePath}
\title{Define File to extract Path}
\usage{
set_extractFilePath(extractFolder, fileName)
}
\arguments{
\item{extractFolder}{path to the folder where the file is placed}

\item{fileName}{name of the file}
}
\value{
filePath
}
\description{
constructs the path to the file to extract
}
