% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{set_extractFolder}
\alias{set_extractFolder}
\title{Define Extract Folder Path}
\usage{
set_extractFolder(id, label)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{label}{either input or output}
}
\value{
extractFolder
}
\description{
constructs the path to the folder where to extract files
}
