% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{showname}
\alias{showname}
\title{return element based on length of input}
\usage{
showname(x)
}
\arguments{
\item{x}{list, NULL or string}
}
\value{
string
}
\description{
Returns the element name of the input x if x has legth > 1, the elemnt x itself if it has length = 1, and "Not available" if if x has legth = 0
}
