% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{showNotification}
\alias{showNotification}
\title{Show or remove a notification}
\usage{
showNotification(ui, type = c("default", "message", "warning", "error"), ...)
}
\arguments{
\item{ui}{see \link{shiny::showNotification}}

\item{type}{see \link{shiny::showNotification}}

\item{...}{other arguments to \link{shiny::showNotification}}
}
\description{
Modified version of the default \link{shiny::showNotifcation}
}
