% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R
\name{sidebar_button}
\alias{sidebar_button}
\title{sidebar_button}
\usage{
sidebar_button(ID, Label = NULL, Icon = NULL, Block = TRUE,
  Style = "btn btn-primary")
}
\description{
function containing the sidebar UI
}
