% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAna_server.R, R/singleAna_ui.R
\name{singleAna}
\alias{singleAna}
\alias{singleAnaUI}
\title{singleAna}
\usage{
singleAna(input, output, session, active = reactive(TRUE),
  logMessage = message, selectAnaID = reactive(""),
  selectPortfolioID = reactive(""), preselPanel = reactive(1))

singleAnaUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{selectAnaID}{id of selected analysis as returned from \link{visualizationSBR}}

\item{selectPortfolioID}{portfolio id of selected analysis as returned from \link{visualizationSBR}}

\item{preselPanel}{selectedstep to visualize as returned from either
\link{visualizationSBR}, \link{visualizationCBR} or \link{visualizationBBR}}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

anaID  selected analysis ID

List of tags.
}
\description{
Server logic to define an analysis

UI/Viewt o define an analysis
}
