% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleAnaWorkflowSteps.R
\name{singleAnaWorkflowSteps}
\alias{singleAnaWorkflowSteps}
\title{singleAnaWorkflowSteps}
\usage{
singleAnaWorkflowSteps(input, output, session)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{id}{Namespace ID for the module.}
}
\value{
smth.
}
\description{
singleAnaWorkflowSteps
}
