% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleProgrammeWorkflowSteps.R
\name{singleAnaWorkflowStepsUI}
\alias{singleAnaWorkflowStepsUI}
\title{singleAnaWorkflowStepsUI}
\usage{
singleAnaWorkflowStepsUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
List of tags.
}
\description{
singleAnaWorkflowStepsUI
}
