% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleProgrammeWorkflowSteps.R
\name{singleProgrammeWorkflowSteps}
\alias{singleProgrammeWorkflowSteps}
\title{singleProgrammeWorkflowSteps}
\usage{
singleProgrammeWorkflowSteps(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\description{
singleProgrammeWorkflowSteps
}
