% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleProgrammeWorkflowSteps.R
\name{singleProgrammeWorkflowStepsUI}
\alias{singleProgrammeWorkflowStepsUI}
\title{singleProgrammeWorkflowStepsUI}
\usage{
singleProgrammeWorkflowStepsUI(id)
}
\arguments{
\item{id}{id}
}
\description{
singleProgrammeWorkflowStepsUI
}
