% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{sliderInputBlock}
\alias{sliderInputBlock}
\title{Function wrapping slider input}
\usage{
sliderInputBlock(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping slider input
}
