% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusdetail_module.R
\name{statusdetailUI}
\alias{statusdetailUI}
\alias{statusdetail}
\title{statusdetailUI}
\usage{
statusdetailUI(id)

statusdetail(input, output, session, analysisID)
}
\arguments{
\item{analysisID}{Selected analysis id.}
}
\value{
List of tags.
}
\description{
UI/View for status detail of files of an analysis.

Server logic for status detail of files of an analysis.
}
