% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{step1_chooseProgrammeUI}
\alias{step1_chooseProgrammeUI}
\alias{step1_chooseProgramme}
\title{step1_chooseProgrammeUI}
\usage{
step1_chooseProgrammeUI(id)

step1_chooseProgramme(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive(""))
}
\arguments{
\item{id}{Shiny module id.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{programmeDefinitionSingle()}, list of reactives.
}
\description{
UI/View for the step1_chooseProgramme.

Server logic to step1_chooseProgramme.
}
