% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel.R
\name{step2_chooseModelUI}
\alias{step2_chooseModelUI}
\alias{step2_chooseModel}
\title{step2_chooseModel UI}
\usage{
step2_chooseModelUI(id)

step2_chooseModel(input, output, session, dbSettings, apiSettings,
  active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive({     "" }),
  selectprogOasisID = reactive({     "" }), progName = reactive({    
  "" }), progStatus = reactive({     "" }),
  DPProgData = reactive(NULL))
}
\arguments{
\item{id}{module id}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags

For \code{programmeDefinitionSingle()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View for the step2_chooseModel

Server logic to step2_chooseModel
}
