% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{step3_configureOutputUI}
\alias{step3_configureOutputUI}
\alias{step3_configureOutput}
\title{step3_configureOutput ui}
\usage{
step3_configureOutputUI(id)

step3_configureOutput(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive(""),
  selectprogOasisID = reactive(""), progOasisName = reactive(""),
  progOasisStatus = reactive(""), POData_rowselected = reactive(NULL))
}
\arguments{
\item{id}{module id}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags

For \code{programmeDefinitionSingle()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View for the step3_configureOutput

Server logic to step3_configureOutput
}
