% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{summarytabUI}
\alias{summarytabUI}
\title{summarytab_ui
Run Summary UI}
\usage{
summarytabUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
Summary elements of a Run
}
