% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_settings.R
\docType{data}
\name{tbl_analysesDataNames}
\alias{tbl_analysesDataNames}
\title{tbl_analysesDataNames}
\format{Named \code{list} of tbl_analysesData column names.}
\usage{
tbl_analysesDataNames
}
\description{
Creating List for col names of Model Runs Table
}
\keyword{datasets}
