% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{testFlamingoServer}
\alias{testFlamingoServer}
\title{Flamingo API Server health check}
\usage{
testFlamingoServer(apiSettings, ...)
}
\arguments{
\item{apiSettings}{settings object as returned by \link{flamingoServer}}

\item{...}{further arguments to \link{GET}}
}
\value{
\code{200} if the health check succeeds; otherwise an exception is
raised.
}
\description{
perform a health check on the Flamingo API Server
}
