% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{untar_file}
\alias{untar_file}
\title{untar file}
\usage{
untar_file(tarfile, file_name, destdir = tempdir(), nrows = Inf,
  to_strip = "")
}
\arguments{
\item{tarfile}{path to tar bundle.}

\item{file_name}{name of file to untar.}

\item{destdir}{path where to extract file.}

\item{nrows}{number of rows to read. Default Inf indicates full content.}

\item{to_strip}{character vector of initial path to strip relative to the tar bundle content.}
}
\value{
content of file_name
}
\description{
Returns list of files in a tar bundle.
}
