% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadedinputs_module.R
\name{uploadedinputsUI}
\alias{uploadedinputsUI}
\alias{uploadedinputs}
\title{uploadedinputsUI}
\usage{
uploadedinputsUI(id)

uploadedinputs(input, output, session, analysisID)
}
\arguments{
\item{analysisID}{Selected analysis id.}
}
\value{
List of tags.
}
\description{
UI/View for uploaded inputs of an analysis.

Server logic for uploaded inputs of an analysis.
}
