% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUserAdmin_server.R,
%   R/defineUserAdmin_ui.R
\name{userAdminDefinition}
\alias{userAdminDefinition}
\alias{userAdminDefinitionUI}
\title{userAdminDefinition}
\usage{
userAdminDefinition(input, output, session, dbSettings, userId,
  active = reactive(TRUE))

userAdminDefinitionUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{id}{Shiny module id.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{userAdminDefinition()}, list of reactives.

List of tags.
}
\description{
Server logic for accessing the Company User List for Flamingo
in association with OASIS LMF.

UI/View for accessing the Company User List for Flamingo
in association with OASIS LMF.
}
