% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewDTTable_module.R
\name{ViewDTTableUI}
\alias{ViewDTTableUI}
\alias{ViewDTTable}
\title{ViewDTTableUI}
\usage{
ViewDTTableUI(id)

ViewDTTable(input, output, session)
}
\value{
List of tags.
}
\description{
UI/View to view  files.

Server logic to view DT table.
}
