% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewFilesInTable_module.R
\name{ViewFilesInTableUI}
\alias{ViewFilesInTableUI}
\alias{ViewFilesInTable}
\title{ViewFilesInTableUI}
\usage{
ViewFilesInTableUI(id, includechkbox = FALSE)

ViewFilesInTable(input, output, session, logMessage = message,
  tbl_filesListData, param = NULL, file_column = "files",
  folderpath = "_output/output/", includechkbox = FALSE)
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{logMessage}{function that will be passed info messages.}

\item{tbl_filesListData}{dataframe of output files.}

\item{param}{id to be used}

\item{file_column}{name of the column containing filename. Default "files"}

\item{folderpath}{path to files. Can be "_output/output/" or "_inputs/"; default output path.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View to view  files.

Server logic to view files.
}
