% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModuleUI}
\alias{ViewFilesModuleUI}
\title{UI for ViewFilesModule}
\usage{
ViewFilesModuleUI(id, includechkbox = FALSE)
}
\arguments{
\item{id}{module id}
}
\description{
UI for ViewFilesModule
}
