% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModule}
\alias{ViewFilesModule}
\title{Server logic for ViewFilesModule}
\usage{
ViewFilesModule(input, output, session, logMessage = message,
  filesListData, includemrows = FALSE, includechkbox = FALSE)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{logMessage}{function that will be passed info messages}

\item{filesListData}{table of output files for a given runID}
}
\value{
list of reactives:
}
\description{
Server logic to view  files
}
