% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_ui.R
\name{visualizationBBRUI}
\alias{visualizationBBRUI}
\title{visualizationBBR_ui
visualization Batch Browse Run UI}
\usage{
visualizationBBRUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
UI/View for the batch process run page
}
