% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R,
%   R/visualizationCBR_ui.R
\name{visualizationCBR}
\alias{visualizationCBR}
\alias{visualizationCBRUI}
\title{Compare Runs Definition Module}
\usage{
visualizationCBR(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

visualizationCBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{Shiny module id.}
}
\value{
For \code{visualizationCBR()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic to compare runs

UI/View for compare runs run page
}
