% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_ui.R
\name{visualizationCBRUI}
\alias{visualizationCBRUI}
\title{visualizationCBR_ui
visualization Compare Browse Run UI}
\usage{
visualizationCBRUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
UI/View for the compare process runs page
}
