% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{visualizationSBRUI}
\alias{visualizationSBRUI}
\title{visualizationSBR_ui
visualization Single Browse Run UI}
\usage{
visualizationSBRUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
UI/View for the process run page
}
