% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hub_helper_funcs.R
\name{write_file}
\alias{write_file}
\title{write file}
\usage{
write_file(data, dataset_identifier = NULL, destdir = tempdir(),
  file_towrite = NULL)
}
\arguments{
\item{data}{object to write.}

\item{dataset_identifier}{name and relative path of file to write}

\item{destdir}{path where to write file.}
}
\description{
Writes ojbect in the correct format.
}
