% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAccount_server.R, R/defineAccount_ui.R
\name{accountDefinition}
\alias{accountDefinition}
\alias{accountDefinitionUI}
\title{Account Definition Module}
\usage{
accountDefinition(input, output, session, dbSettings,
  active = reactive(TRUE))

accountDefinitionUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{dbSettings}{Setting object as returned by e.g. [flamingoDB()].}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

Empty list.

List of tags.
}
\description{
Server logic to define an account.

UI/View to define an account.
}
