% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAccount_server.R, R/defineAccount_ui.R
\name{accountDefinition}
\alias{accountDefinition}
\alias{accountDefinitionUI}
\title{Account Definition Module}
\usage{
accountDefinition(input, output, session, dbSettings, active = reactive(TRUE))

accountDefinitionUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{shiny module id}
}
\value{
empty list

list of tags
}
\description{
Server logic to define an account.

UI/View to define an account.
}
