% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAccount_server.R, R/defineAccount_ui.R
\name{accountDefinition}
\alias{accountDefinition}
\alias{accountDefinitionUI}
\title{Account Definition Module}
\usage{
accountDefinition(input, output, session, dbSettings,
  active = reactive(TRUE))

accountDefinitionUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{logMessage}{function that will be passed info messages.}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

Empty list.

List of tags.
}
\description{
Server logic to define an account.

UI/View to define an account.
}
