% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAccount_ui.R
\name{accountDefinitionUI}
\alias{accountDefinitionUI}
\title{accountDefinitionUI}
\usage{
accountDefinitionUI(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\value{
List of tags.
}
\description{
UI/View to define an account.
}
