% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anainputs_module.R
\name{anainputsUI}
\alias{anainputsUI}
\alias{anainputs}
\title{anainputsUI}
\usage{
anainputsUI(id)

anainputs(input, output, session, analysisID, portfolioID,
  counter = NULL, active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis ID.}

\item{portfolioID}{Selected portfolio ID.}

\item{counter}{Reactive value storing actionButton status.}
}
\value{
List of tags.
}
\description{
UI/View for inputs of an analysis.

Server logic for inputs of an analysis.
}
