% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{api_access_token}
\alias{api_access_token}
\title{Post access token}
\usage{
api_access_token(user, pwd)
}
\arguments{
\item{user}{Username.}

\item{pwd}{Password.}
}
\value{
List with API return status and response containing the new token.
}
\description{
Fetches a new access token from a username and password.
}
