% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{api_get_analyses}
\alias{api_get_analyses}
\title{Get analyses}
\usage{
api_get_analyses(name = "")
}
\arguments{
\item{name}{the name of the analysis. Default empty string returns all analyses}
}
\value{
previously posted analyses.
}
\description{
Returns a list of analysis objects.
}
