% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{api_get_models_id}
\alias{api_get_models_id}
\title{Get models id}
\usage{
api_get_models_id(id)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}
}
\value{
previously posted models id.
}
\description{
Returns the specific model entry.
}
