% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{api_get_portfolios}
\alias{api_get_portfolios}
\title{Get portfolios}
\usage{
api_get_portfolios(name = "")
}
\arguments{
\item{name}{Name of the portfolio. Default empty string returns all portfolios.}
}
\value{
Previously posted portfolios.
}
\description{
Returns a list of portfolio objects.
}
