% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses.R
\name{api_post_analyses}
\alias{api_post_analyses}
\title{Post analyses}
\usage{
api_post_analyses(name, portfolio, model)
}
\arguments{
\item{name}{the name of the analysis.}

\item{portfolio}{the id of the portfolio.}

\item{model}{the id of the model.}
}
\value{
the posted analysis.
}
\description{
Creates an analysis based on the input data
}
