% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_generate_inputs}
\alias{api_post_analyses_generate_inputs}
\title{Post analysis input generation}
\usage{
api_post_analyses_generate_inputs(id)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}
}
\value{
The posted analysis input generation.
}
\description{
Sets the analysis generate_inputs contents.
}
