% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{api_post_analyses_run_traceback_file}
\alias{api_post_analyses_run_traceback_file}
\title{Post analysis run_traceback file}
\usage{
api_post_analyses_run_traceback_file(id, filepath_run_traceback)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_run_traceback}{path to the run_traceback file.}
}
\value{
the posted analysis run_traceback file.
}
\description{
Sets the analysis run_traceback_file contents.
}
