% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_post_portfolios_accounts_file}
\alias{api_post_portfolios_accounts_file}
\title{Post portfolios accounts file}
\usage{
api_post_portfolios_accounts_file(id, filepath_accounts)
}
\arguments{
\item{id}{A unique integer value identifying this portfolio.}

\item{filepath_accounts}{Path to accounts file.}
}
\value{
The posted portfolio accounts file.
}
\description{
Sets the portfolios accounts_file contents.
}
