% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_post_portfolios_location_file}
\alias{api_post_portfolios_location_file}
\title{Post portfolios location file}
\usage{
api_post_portfolios_location_file(id, filepath_location)
}
\arguments{
\item{id}{a unique integer value identifying this analysis.}

\item{filepath_location}{path to the location file.}
}
\value{
the posted portfolio location file.
}
\description{
Sets the portfolios location_file contents.
}
