% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios_id.R
\name{api_post_portfolios_reinsurance_source_file}
\alias{api_post_portfolios_reinsurance_source_file}
\title{Post portfolios reinsurance source file}
\usage{
api_post_portfolios_reinsurance_source_file(id,
  filepath_reinsurance_source)
}
\arguments{
\item{id}{A unique integer value identifying this analysis.}

\item{filepath_reinsurance_source}{Path to reinsurance source file.}
}
\value{
the posted portfolio reinsurance source file.
}
\description{
Sets the portfolios reinsurance_source contents.
}
