% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_setup.R
\name{api_refresh_token}
\alias{api_refresh_token}
\title{Post refresh token}
\usage{
api_refresh_token()
}
\value{
List with API return status and response containing the new token.
}
\description{
Fetches a new access token from an existing refresh token.
}
