% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{buildDbQuery}
\alias{buildDbQuery}
\title{buildDbQuery}
\usage{
buildDbQuery(dbFunc, ..., dboPrefix = TRUE, squareBrackets = FALSE)
}
\arguments{
\item{dbFunc}{Database function to call.}

\item{...}{Arguments to \code{dbFunc}. \code{character()} args are quoted with single
quotes (').}

\item{dboPrefix}{Prefix \code{dbFunc} with \code{"dbo."}.}

\item{squareBrackets}{Surround the argument list with square brackets.}
}
\description{
Constructs query statements that can be executed with
\code{\link[DBI:dbExecute]{DBI::dbExecute()}}.
}
