% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoIncrementalPanel.R
\name{callIncrementalPanelModules}
\alias{callIncrementalPanelModules}
\title{callIncrementalPanelModules}
\usage{
callIncrementalPanelModules(IDs, ID_0, contentIDs, contentUI, ...,
  headings = NULL, collapsible = FALSE, show = TRUE, ns = identity)
}
\arguments{
\item{IDs}{IDs.}

\item{ID_0}{ID_0.}

\item{contentIDs}{content of the IDs.}

\item{contentUI}{content of the UI.}

\item{...}{Additional parameters.}

\item{collapsible}{FALSE.}

\item{show}{TRUE.}
}
\value{
smth.
}
\description{
callIncrementalPanelModules
}
