% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{collapseButton}
\alias{collapseButton}
\title{collapseButton}
\usage{
collapseButton(id, id_collapse, ..., width = NULL, collapsed = FALSE)
}
\arguments{
\item{id}{id}

\item{id_collapse}{id to collapse}

\item{...}{additional paramters}

\item{width}{NULL}

\item{collapsed}{FALSE}
}
\description{
collapseButton
}
