% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{collapseButton}
\alias{collapseButton}
\title{collapseButton}
\usage{
collapseButton(id, id_collapse, ..., width = NULL, collapsed = FALSE)
}
\arguments{
\item{id}{Shiny module id.}

\item{...}{Additional paramters.}

\item{width}{Width of the column, a reactive value for \code{dynamicColumn()},
non-reactive (or \link[shiny:isolate]{isolated}) for \code{dynamicColumnUI()}.}

\item{collapsed}{FALSE.}
}
\value{
Collapsed button.
}
\description{
collapseButton
}
