% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompany_server.R, R/defineCompany_ui.R
\name{companyDefinition}
\alias{companyDefinition}
\alias{companyDefinitionUI}
\title{Company Definition Module}
\usage{
companyDefinition(input, output, session, dbSettings, userId,
  active = reactive(TRUE), logMessage = message)

companyDefinitionUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{userId}{reactive expression yielding user id}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{id}{account id}
}
\value{
empty list

list of tags
}
\description{
Server logic to define a company

UI/View to define an account.
}
