% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompany_server.R
\name{companyDefinition}
\alias{companyDefinition}
\title{Company Definition Module}
\usage{
companyDefinition(input, output, session, dbSettings,
  active = reactive(TRUE), logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
empty list
}
\description{
Server logic to define a company
}
