% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCompany_server.R, R/defineCompany_ui.R
\name{companyDefinition}
\alias{companyDefinition}
\alias{companyDefinitionUI}
\title{companyDefinition}
\usage{
companyDefinition(input, output, session, dbSettings,
  active = reactive(TRUE), logMessage = message)

companyDefinitionUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

Empty list.

List of tags.
}
\description{
Server logic to define a company.

UI/View to define an account.
}
