% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{configureAdvancedGUL}
\alias{configureAdvancedGUL}
\title{configureAdvancedGUL}
\usage{
configureAdvancedGUL(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\description{
Function wrapping sub-panel to define outputs advanced configuration GUL.
}
