% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_analyses_id.R
\name{construct_tbl_analyses_settings}
\alias{construct_tbl_analyses_settings}
\title{Construct analysis settings list}
\usage{
construct_tbl_analyses_settings(source_tag, prog_id, number_of_samples,
  module_supplier_id, model_version_id, event_occurrence_file_id,
  use_random_number_file = FALSE, event_set, peril_wind, demand_surge,
  peril_quake, peril_flood, peril_surge, leakage_factor, gul_threshold,
  exposure_location = "L", outputsGUL, outputsIL, outputsRI,
  chkinputsummaryoption, gul_output, il_output, ri_output,
  return_period_file, analysis_tag, uniqueItems = FALSE, id = 1)
}
\arguments{
\item{source_tag}{model associated to the analysis.}

\item{prog_id}{portfolio id?}

\item{number_of_samples}{user input number of samples}

\item{module_supplier_id}{identification of model supplied}

\item{model_version_id}{identification of model version}

\item{event_occurrence_file_id}{user input Event Occurrence Set}

\item{use_random_number_file}{use_random_number_file}

\item{event_set}{user input Event Set}

\item{peril_wind}{peril wind}

\item{demand_surge}{demand surge}

\item{peril_quake}{peril quake}

\item{peril_flood}{peril flood}

\item{peril_surge}{peril surge}

\item{leakage_factor}{user input Leakage factor}

\item{gul_threshold}{user input gul threshold}

\item{exposure_location}{exposure_location}

\item{outputsGUL}{output config GUL}

\item{outputsIL}{output config IL}

\item{outputsRI}{output config RI}

\item{chkinputsummaryoption}{logical to include summaries}

\item{gul_output}{logical for GUL}

\item{il_output}{logical for IL}

\item{ri_output}{logical for RI}

\item{return_period_file}{param = TRUE}

\item{analysis_tag}{analysis id?}

\item{uniqueItems}{uniqueItems}

\item{id}{id}
}
\value{
list of analysis settings.
}
\description{
Constructs a list of analysis settings
}
\details{
Construct analysis_settingsList
}
