% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createFileRecord}
\alias{createFileRecord}
\title{Create file record}
\usage{
createFileRecord(dbSettings, fileName, fileDesc, fileType, locPathUnix,
  BFEUserID, ResourceTable, ResourceKey)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{fileName}{file name}

\item{fileDesc}{file descriptor}

\item{fileType}{file type}

\item{locPathUnix}{locPathUnix}

\item{BFEUserID}{BFEUserID}

\item{ResourceTable}{ResourceTable}

\item{ResourceKey}{ResourceKey}
}
\value{
file record id (integer)
}
\description{
document args better
}
