% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createFileRecord}
\alias{createFileRecord}
\title{createFileRecord}
\usage{
createFileRecord(dbSettings, fileName, fileDesc, fileType, locPathUnix,
  BFEUserID, ResourceTable, ResourceKey)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{fileName}{File name.}

\item{fileDesc}{File describer.}

\item{fileType}{File type.}

\item{locPathUnix}{locPathUnix.}

\item{BFEUserID}{BFEUserID.}

\item{ResourceTable}{ResourceTable.}

\item{ResourceKey}{ResourceKey.}
}
\value{
File record id (integer).
}
\description{
Document args better.
}
