% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlainMap_module.R
\name{createHazardMapUI}
\alias{createHazardMapUI}
\alias{createHazardMap}
\title{createHazardMap}
\usage{
createHazardMapUI(id)

createHazardMap(input, output, session, file)
}
\arguments{
\item{file}{File to plot as map.}
}
\value{
Leaflet map.
}
\description{
Creates a hazard map using leaflet.

Creates a hazard map using leaflet.
}
