% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createModelResource}
\alias{createModelResource}
\title{Create Model Resource}
\usage{
createModelResource(dbSettings, modresname, restype, oasissys, modelid,
  modresval)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{modresname}{\code{character()}; name of the model resource}

\item{restype}{resource type}

\item{oasissys}{TODO document}

\item{modelid}{model id}

\item{modresval}{TODO document}
}
\value{
id
}
\description{
TODO document
}
