% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createModelResource}
\alias{createModelResource}
\title{createModelResource}
\usage{
createModelResource(dbSettings, modresname, restype, oasissys, modelid,
  modresval)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{modresname}{\code{character()}; name of the model resource.}

\item{restype}{Resource type.}

\item{oasissys}{TODO document.}

\item{modelid}{Model id.}

\item{modresval}{TODO document.}
}
\value{
id.
}
\description{
Create Model Resource.
}
