% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{createPlainMap}
\alias{createPlainMap}
\title{Create Plain Map}
\usage{
createPlainMap(fileName)
}
\arguments{
\item{fileName}{full file path to marker data}
}
\description{
creates a plain map using leaflet
}
