% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlainMap_module.R
\name{createPlainMapUI}
\alias{createPlainMapUI}
\alias{createPlainMap}
\title{createPlainMap_ui}
\usage{
createPlainMapUI(id)

createPlainMap(input, output, session, file)
}
\arguments{
\item{file}{File to plot as map.}
}
\value{
Leaflet map.
}
\description{
Creates a plain map using leaflet.

Creates a plain map using leaflet.
}
