% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{createProgOasis}
\alias{createProgOasis}
\title{createProgOasis}
\usage{
createProgOasis(dbSettings, progId, modelId, transformId)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{progId}{Programme id.}

\item{modelId}{Model id.}

\item{transformId}{Transform id.}
}
\value{
Prog Oasis id.
}
\description{
Create Prog Oasis.
}
