% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahub.R
\docType{class}
\name{DataHub}
\alias{DataHub}
\title{DataHub}
\format{\code{\link{R6Class}} object.}
\usage{
DataHub
}
\value{
Object of \code{\link{R6Class}} with methods for Oasis files lists and Datatasets explorations.
}
\description{
R6 Class for OasisUI Data Hub.
}
\section{Arguments}{

\describe{
\item{id}{Portfolio Id | Model Id | Analysis Id}
\item{type}{Identifier for Analysis Id: inputs | outputs}
\item{dataset_identifier}{Identifier:  file name for Analysis Id |  api call for Portfolio Id and Model Id}
}
}

\section{Privates}{

\describe{
\item{user}{string identifying the user.}
\item{destdir = tempdir()}{user specific destdir.}
}
}

\section{Usage}{

\preformatted{data_hub <- DataHub$new()
data_hub$get_pf_data_list(id)
}
}

\keyword{datasets}
