% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineID.R
\name{defineID}
\alias{defineID}
\title{define oneID Module}
\usage{
defineID(input, output, session, dbSettings, userId,
  runIdList = reactive(c(-1)), preselectedRunId = reactive(-1),
  logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{runIdList}{list of runs and their status}

\item{preselectedRunId}{reactive string expression for reselected run id from landingpage}

\item{logMessage}{function that will be passed info messages}
}
\value{
reactive for runID selected
}
\description{
Server logic for defining one run ID
}
