% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineID.R
\name{defineIDUI}
\alias{defineIDUI}
\alias{defineID}
\title{defineIDUI}
\usage{
defineIDUI(id, w, batch = FALSE)

defineID(input, output, session, dbSettings, userId,
  preselectedRunId = reactive(-1), batch = FALSE,
  logMessage = message)
}
\arguments{
\item{id}{module id}

\item{w}{width of the coulmn.}

\item{batch}{flag indicating if it is a batch or a simple run.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{preselectedRunId}{reactive string expression for reselected run id from \link{landingpage}}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{batch}{Flag indicating if it is a batch or a simple run.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

selectRunID reactive for runID selected.
}
\description{
UI/View for defining one run ID

Server logic for defining one run ID.
}
