% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineID.R
\name{defineIDUI}
\alias{defineIDUI}
\alias{defineID}
\title{defineIDUI}
\usage{
defineIDUI(id, w, batch = FALSE)

defineID(input, output, session, preselAnaId = reactive(-1),
  anaID = reactive(-1), batch = FALSE, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{w}{width of the coulmn.}

\item{batch}{flag indicating if it is a batch or a simple run.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingPage}.}

\item{anaID}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{batch}{Flag indicating if it is a batch or a simple analysis.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

selectAnaID reactive for anaID selected.
}
\description{
UI/View for defining one analysis ID

Server logic for defining one analysis ID.
}
