% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{deleteModelResource}
\alias{deleteModelResource}
\title{deleteModelResource}
\usage{
deleteModelResource(dbSettings, modresid)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{modelid}{Model id.}
}
\value{
Model resource id.
}
\description{
Delete Model Resource.
}
