% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{deleteModelResource}
\alias{deleteModelResource}
\title{Delete Model Resource}
\usage{
deleteModelResource(dbSettings, modresid)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{modresid}{model resource id}
}
\value{
model resource id
}
\description{
Delete Model Resource
}
