% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{executeDbQuery}
\alias{executeDbQuery}
\title{executeDbQuery}
\usage{
executeDbQuery(dbSettings, statement, simplify = FALSE,
  verbose = FALSE, logMessage = message, logError = logMessage, ...)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{statement}{Query statement to execute, e.g. as returned by
\code{\link[=buildDbQuery]{buildDbQuery()}}.}

\item{simplify}{If \code{TRUE}, a vector will be returned instead of a
\code{data.frame} only a single column is returned.}

\item{verbose}{Print query statement.}

\item{logMessage}{Function to call with log message.}

\item{logError}{Function to call with error log message.}

\item{...}{Further arguments to \code{\link[=dbGetQuery]{dbGetQuery()}}.}
}
\value{
Result of the query as a \code{data.frame} or a vector if \code{simplify} is
set to \code{TRUE}. \code{NULL} if no result can be returned.
}
\description{
Opens a new connection to the database, executes the given query
statement and then closes the connection again.
}
