% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurevalidation_module.R
\name{exposurevalidationUI}
\alias{exposurevalidationUI}
\alias{exposurevalidation}
\title{exposurevalidationUI}
\usage{
exposurevalidationUI(id)

exposurevalidation(input, output, session, analysisID = "",
  portfolioID = "", counter = NULL, active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis id.}

\item{portfolioID}{selected portfolio ID.}

\item{counter}{Reactive value storing actionButton status.}
}
\value{
List of tags.
}
\description{
UI/View for exposure validation of an analysis.

Server logic for exposure validation of an analysis.
}
