% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposurevalidation_module.R
\name{exposurevalidationUI}
\alias{exposurevalidationUI}
\alias{exposurevalidation}
\title{exposurevalidationUI}
\usage{
exposurevalidationUI(id)

exposurevalidation(input, output, session)
}
\value{
List of tags.
}
\description{
UI/View for uploaded inputs of an analysis.

Server logic for uploaded inputs of an analysis.
}
