% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileViewer_server.R, R/fileViewer_ui.R
\name{fileViewer}
\alias{fileViewer}
\alias{fileViewerUI}
\title{File Viewer Module}
\usage{
fileViewer(input, output, session, dbSettings, userId,
  preselRunId = reactive(-1), active, logMessage = message)

fileViewerUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{preselRunId}{Reactive string expression for reselected run id from landingpage.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}
}
\value{
Empty list.
}
\description{
Server logic to view files.

UI/View to view files.
}
