% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileViewer_server.R, R/fileViewer_ui.R
\name{fileViewer}
\alias{fileViewer}
\alias{fileViewerUI}
\title{File Viewer Module}
\usage{
fileViewer(input, output, session, dbSettings, userId,
  preselRunId = reactive(-1), active, logMessage = message)

fileViewerUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{userId}{user id}

\item{preselRunId}{preselected run id}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
empty list
}
\description{
Server logic to view files

UI/View to view files
}
