% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoCheckboxButton}
\alias{flamingoCheckboxButton}
\title{Flamingo action button look like a checkbox}
\usage{
flamingoCheckboxButton(inputId, label, icon = NULL, width = NULL,
  class = c("btn"), ...)
}
\arguments{
\item{class}{HTML class attribute}

\item{...}{arguments to \code{\link[shiny:actionButton]{shiny::actionButton()}}}
}
\description{
Modified version of the default \code{\link[shiny:actionButton]{shiny::actionButton()}}
}
