% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{flamingoDB}
\alias{flamingoDB}
\title{flamingoDB}
\usage{
flamingoDB(server, port, database, uid, pwd, driver = "{FreeTDS}")
}
\arguments{
\item{server}{Host name.}

\item{port}{Host port.}

\item{database}{Database name.}

\item{pwd}{User pass.}

\item{driver}{Connection driver.}
}
\value{
A database setttings object (list).
}
\description{
Creates a database settings object which can then be used
to create new connections to the Flamingo Database.
}
