% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{flamingoDBLogin}
\alias{flamingoDBLogin}
\title{flamingoDBLogin}
\usage{
flamingoDBLogin(dbSettings, uid, pwd)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{uid}{user name}

\item{pwd}{user password}
}
\value{
User id if succesful, -1 otherwise.
}
\description{
Construct and execute a login query for the given credentials.
}
