% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoIncrementalPanel.R
\name{flamingoIncrementalPanel}
\alias{flamingoIncrementalPanel}
\title{flamingoIncrementalPanel}
\usage{
flamingoIncrementalPanel(input, output, session, panels_state,
  new_content_IDs, new_content_fun, ..., new_headings = NULL,
  collapsible = FALSE, show = TRUE)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{panels_state}{state of panel}

\item{new_content_IDs}{ID with new content}

\item{new_content_fun}{function with new content}

\item{new_headings}{new heading}
}
\description{
flamingoIncrementalPanel
}
