% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flamingoModule}
\alias{flamingoModule}
\title{Flamingo Module Template}
\usage{
flamingoModule(input, output, session, active, dbSettings, apiSettings,
  logMessage, logError)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{logMessage}{Function that will be passed info messages.}

\item{logError}{Function that will be passed error messages.}
}
\description{
Flamingo Module Template
}
