% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R, R/step2_chooseModel.R
\name{flamingoModuleUI}
\alias{flamingoModuleUI}
\alias{panelAssociateModel}
\title{flamingoModuleUI}
\usage{
flamingoModuleUI(id)

panelAssociateModel(id)
}
\arguments{
\item{id}{Namespace ID for the module.}
}
\value{
No value.
}
\description{
Flamingo Module Template UI.

Function wrapping panel to associate model.
}
