% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{flamingoPanel}
\alias{flamingoPanel}
\title{flamingoPanel}
\usage{
flamingoPanel(id, ..., heading = NULL, footer = NULL,
  status = "default", collapsible = FALSE, show = TRUE)
}
\arguments{
\item{id}{id}

\item{...}{additional parameters}

\item{heading}{NULL by default}

\item{footer}{NULL by default}

\item{status}{default}

\item{collapsible}{FALSE}

\item{show}{TRUE}
}
\description{
flamingoPanel
}
\details{
adapted from shinyWidgets::panel() including IDs and collapsible
(inspired by shinyBS::bsCollapsePanel())
}
