% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{flamingoPanelHeadingOutput}
\alias{flamingoPanelHeadingOutput}
\title{flamingoPanelHeadingOutput}
\usage{
flamingoPanelHeadingOutput(outputId, ...)
}
\arguments{
\item{outputId}{outputId}

\item{...}{additional parameters}
}
\description{
flamingoPanelHeadingOutput
}
