% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{flamingoPanelHeadingOutput}
\alias{flamingoPanelHeadingOutput}
\title{Panel Heading UI Output}
\usage{
flamingoPanelHeadingOutput(outputId, ...)
}
\arguments{
\item{outputId}{ID of the dynamical panel heading.}

\item{...}{Other arguments to pass to the container tag function. This is
useful for providing additional classes for the tag.}
}
\description{
Panel heading UI placeholder to be rendered dynamically via
\code{\link[=renderFlamingoPanelHeading]{renderFlamingoPanelHeading()}}.
}
\seealso{
\code{\link[=renderFlamingoPanelHeading]{renderFlamingoPanelHeading()}}
}
