% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{flamingoServer}
\alias{flamingoServer}
\title{Create a settings object for the Flamingo API Server}
\usage{
flamingoServer(host, port, scheme = c("http", "https"))
}
\arguments{
\item{port}{host port}

\item{scheme}{communication scheme}

\item{server}{host name}
}
\value{
a flamingo server settings object (HTTP by default)
}
\description{
creates a settings object which can then be used to connect
to the Flamingo API Server.
}
