% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoTable_module.R
\name{flamingoTableUI}
\alias{flamingoTableUI}
\alias{flamingoTable}
\title{flamingoTableUI}
\usage{
flamingoTableUI(id)

flamingoTable(input, output, session, data, selection = "none",
  escape = TRUE, scrollX = FALSE, filter = FALSE, rownames = FALSE,
  colnames = TRUE, preselRow = reactive({     NULL }),
  maxrowsperpage = 10, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{data}{dataframe to show in table.}

\item{selection}{param of datatable, default"none".}

\item{escape}{param of datatable, default TRUE.}

\item{scrollX}{param of datatable, default FALSE.}

\item{filter}{param of datatable, default FALSE.}

\item{rownames}{param of datatable, default FALSE.}

\item{colnames}{param of datatable, default TRUE.}

\item{preselRow}{reactive of preselected row default reactive({NULL}).}

\item{maxrowsperpage}{param of datatable, default 10.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

rows_selected reactive of selected rows as returned from datatable.

rows_current reactive of current rows as returned from datatable.
}
\description{
Server logic to show the flamingo table output
}
