% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatedinputs_module.R
\name{generatedinputsUI}
\alias{generatedinputsUI}
\alias{generatedinputs}
\title{generatedinputsUI}
\usage{
generatedinputsUI(id)

generatedinputs(input, output, session, analysisID)
}
\arguments{
\item{analysisID}{selected analysis ID.}

\item{param}{AnalysisID}
}
\value{
List of tags.
}
\description{
UI/View for uploaded inputs of an analysis.

Server logic for uploaded inputs of an analysis.
}
