% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getCompanyList}
\alias{getCompanyList}
\title{Get the list of companies from the database}
\usage{
getCompanyList(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}
}
\value{
companies; \code{data.frame} of 5 variables:
\itemize{
		\item \code{Company ID}
		\item \code{Company Name}
   \item \code{Company Domicile}
		\item \code{Company Legal Name}
		\item \code{Company Registration Number}
}
}
\description{
Get the list of companies from the database
}
