% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getDeptData}
\alias{getDeptData}
\title{Get User Department}
\usage{
getDeptData(dbSettings, user)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{user}{user}
}
\description{
queries the database to retrieve department login and password
details for the designated user.
}
