% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getDeptData}
\alias{getDeptData}
\title{getDeptData}
\usage{
getDeptData(dbSettings, userId)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}
}
\description{
Queries the database to retrieve department login and password
details for the designated user.
}
