% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getEventOccurrence}
\alias{getEventOccurrence}
\title{Get the event occurrence list}
\usage{
getEventOccurrence(dbSettings, prgoasisid, simplify = TRUE)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{prgoasisid}{oasis programme id}

\item{simplify}{if \code{TRUE}, a vector will be returned instead of
a \code{data.frame} only a single column is returned.}
}
\value{
event occurrence
}
\description{
Get the event occurrence list
}
