% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getEventSet}
\alias{getEventSet}
\title{Get the event set list}
\usage{
getEventSet(dbSettings, prgoasisid, simplify = TRUE)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{prgoasisid}{oasis programme id}

\item{simplify}{if \code{TRUE}, a vector will be returned instead of a
\code{data.frame} only a single column is returned.}
}
\value{
event set
}
\description{
Get the event set list
}
