% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getEventSet}
\alias{getEventSet}
\title{getEventSet}
\usage{
getEventSet(dbSettings, prgoasisid, simplify = TRUE)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{prgoasisid}{Oasis programme id.}

\item{simplify}{If \code{TRUE}, a vector will be returned instead of a
\code{data.frame} only a single column is returned.}
}
\value{
Event set.
}
\description{
Get the event set list.
}
