% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileLocationPath}
\alias{getFileLocationPath}
\title{Get location path for a given fileType}
\usage{
getFileLocationPath(dbSettings, fileType)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}

\item{fileType}{file type}
}
\value{
path
}
\description{
Get location path for a given fileType
}
