% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileSourceAccountFile}
\alias{getFileSourceAccountFile}
\title{Get Source Account Files}
\usage{
getFileSourceAccountFile(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}
}
\value{
\code{data.frame} of 2 variables:
\itemize{
		\item \code{FileName}
		\item \code{FileId} 
}
}
\description{
Get Source Account Files
}
