% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileSourceAccountFile}
\alias{getFileSourceAccountFile}
\title{getFileSourceAccountFile}
\usage{
getFileSourceAccountFile(dbSettings)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}
}
\value{
res \code{data.frame} of 2 variables:
\itemize{
\item \code{FileName}
\item \code{FileId}
}.
}
\description{
Get Source Account Files.
}
