% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getFileSourceLocationFile}
\alias{getFileSourceLocationFile}
\title{Get File Source Location File}
\usage{
getFileSourceLocationFile(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}
}
\value{
\code{data.frame} of 2 variables:
\itemize{
		\item \code{FileName}
		\item \code{FileId}
}
}
\description{
Get File Source Location File
}
