% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getInboxData}
\alias{getInboxData}
\title{getInboxData}
\usage{
getInboxData(dbSettings, user, ...)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{user}{reactive expression yielding user}

\item{...}{Other arguments to \code{\link[=executeDbQuery]{executeDbQuery()}}.}

\item{dbSettings}{as returned from \link{flamingoDB}}
}
\value{
Inbox; \code{data.frame} of 6 variables:
\itemize{
\item \code{ProgOasisID}
\item \code{AnaID}
\item \code{Run Name}
\item \code{Model}
\item \code{Status}
\item \code{Completed}
}.
}
\description{
Get the model run overview from the database.
}
