% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getInboxData}
\alias{getInboxData}
\title{Get the process run overview from the database}
\usage{
getInboxData(dbSettings, userId, ...)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{...}{other arguments to \code{\link[=executeDbQuery]{executeDbQuery()}}}
}
\value{
inbox; \code{data.frame} of 6 variables:
\itemize{
\item \code{ProgOasisID}
\item \code{RunID}
\item \code{Run Name}
\item \code{Model}
\item \code{Status}
\item \code{Completed}
}
}
\description{
Get the process run overview from the database
}
