% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getOasisSystemId}
\alias{getOasisSystemId}
\title{Get Oasis Systems}
\usage{
getOasisSystemId(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}
}
\value{
oasis systems; \code{data.frame} of 2 variables:
\itemize{
\item \code{OasisSystemID}
\item \code{OasisSystemName}
}
}
\description{
Get Oasis Systems
}
