% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getOasisUsers}
\alias{getOasisUsers}
\title{Get Oasis Users}
\usage{
getOasisUsers(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}
}
\value{
oasis users; \code{data.frame} of 2 variables:
\itemize{
		\item \code{OasisUserID}
		\item \code{OasisUserName}
}
}
\description{
Get Oasis Users
}
