% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getOutputOptions}
\alias{getOutputOptions}
\title{getOutputOptions}
\usage{
getOutputOptions(dbSettings, simplify = TRUE)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{simplify}{logical}
}
\value{
Output options.
}
\description{
Get the list of output presets.
}
