% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcessData}
\alias{getProcessData}
\title{Get Process Data}
\usage{
getProcessData(dbSettings, pruser, prmodel, prprogramme, prworkflow)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{pruser}{process user}

\item{prmodel}{process model}

\item{prprogramme}{process programme}

\item{prworkflow}{process workflow}
}
\value{
process data; \code{data.frame} of 3 variables:
\itemize{
  \item \code{ProgOasisId}
  \item \code{ProgName}
  \item \code{ModelName}
}
}
\description{
Get Process Data
}
