% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcessRun}
\alias{getProcessRun}
\title{getProcessRun}
\usage{
getProcessRun(dbSettings, procid, prstatus)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{procid}{Process id.}

\item{prstatus}{Process status.}
}
\description{
Get Process Run.
}
