% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcessRunDetails}
\alias{getProcessRunDetails}
\title{getProcessRunDetails}
\usage{
getProcessRunDetails(dbSettings, wfid)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{wfid}{Workflow id.}
}
\value{
Log details; \code{data.frame} of 5 variables:
\itemize{
\item \code{Element ID}
\item \code{Element Name}
\item \code{Status}
\item \code{Description}
\item \code{CompletedAt}
}.
}
\description{
Get log details for a run.
}
