% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunDetForFileOutput}
\alias{getProcRunDetForFileOutput}
\title{getProcRunDetForFileOutput}
\usage{
getProcRunDetForFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{processRunId}{Reactive string expression for reselected run id from defineProgramme.}
}
\value{
Process run details.
}
\description{
Get Process Run Details.
}
