% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunDetForFileOutput}
\alias{getProcRunDetForFileOutput}
\title{getProcRunDetForFileOutput}
\usage{
getProcRunDetForFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{processRunId}{identifier selected process run.}
}
\value{
Process run details.
}
\description{
Get Process Run Details.
}
