% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunDetForFileOutput}
\alias{getProcRunDetForFileOutput}
\title{Get Process Run Details}
\usage{
getProcRunDetForFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{processRunId}{process run id}
}
\value{
process run details
}
\description{
Get Process Run Details
}
