% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProcRunParamFileOutput}
\alias{getProcRunParamFileOutput}
\title{Get Process Runtime Param Details}
\usage{
getProcRunParamFileOutput(dbSettings, processRunId)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{processRunId}{process run id}
}
\value{
param details
}
\description{
Get Process Runtime Param Details
}
