% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getProgOasisForProgdata}
\alias{getProgOasisForProgdata}
\title{getProgOasisForProgdata}
\usage{
getProgOasisForProgdata(dbSettings, progId)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{progId}{programme id}
}
\value{
\code{data.frame} of 11 variables:
\itemize{
\item \code{ProgOasisId}
\item \code{ProgName}
\item \code{ModelName}
\item \code{TransformName}
\item \code{SourceFileId}
\item \code{FileID}
\item \code{Status}
\item \code{API1aDateTime}
\item \code{API1bDateTime}
\item \code{API1cDateTime}
\item \code{SessionId}
}.
}
\description{
Get Programme Oasis for a given Programme Id.
}
\details{
Run getProgOasisForProg against the database.
}
