% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getResourceType}
\alias{getResourceType}
\title{Fetch resource type name and ID}
\usage{
getResourceType(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \link{flamingoDB}}
}
\value{
resource types; \code{data.frame} of 3 variables:
\itemize{
		\item \code{ResourceTypeID}
		\item \code{ResourceTypeName}
   \item \code{Source}
}
}
\description{
Fetch resource type name and ID
}
