% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{getSecurityGroups}
\alias{getSecurityGroups}
\title{Get Security Groups}
\usage{
getSecurityGroups(dbSettings)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}
}
\value{
security groups; \code{data.frame} of 2 variables:
\itemize{
		\item \code{Security Group ID}
		\item \code{Security Group Name}
}
}
\description{
Get Security Groups
}
