% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingo_utils.R
\name{getTableOptions}
\alias{getTableOptions}
\title{Get table options.}
\usage{
getTableOptions(scrollX = FALSE, maxrowsperpage = 5, filter = TRUE,
  escape = TRUE)
}
\arguments{
\item{scrollX}{Param to allow scrollX. Default FALSE.}

\item{maxrowsperpage}{Maximum number of rows to display per page. Default 5.}

\item{filter}{Show or hide filter. Default TRUE.}

\item{escape}{Param to avoid escape row. Default TRUE.}
}
\description{
Returns option list for datatable.
}
