% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingpage_module.R
\name{landingPageUI}
\alias{landingPageUI}
\alias{landingPage}
\title{landingPage}
\usage{
landingPageUI(id)

landingPage(input, output, session, user, dbSettings,
  reloadMillis = 10000, logMessage = message,
  active = reactive(TRUE))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{user}{reactive expression yielding user}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{landingPage()}, list of reactives:
\itemize{
		\item{\code{runId}: }{id of selected run or -1 if nothing is selected}
		\item{\code{procId}: }{id of selected process or -1 if nothing is selected}
}
}
\description{
landingPage

Landing Page
}
