% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R, R/landingPage_ui.R
\name{landingPage}
\alias{landingPage}
\alias{landingPageUI}
\title{Landing Page}
\usage{
landingPage(input, output, session, userId, userName, dbSettings,
  reloadMillis = 10000, logMessage = message,
  active = reactive(TRUE))

landingPageUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{userName}{user name as returned by [flamingoDBLogin()]}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{landingPage()}, list of reactives:
\itemize{
		\item{\code{runId}: }{id of selected run or -1 if nothing is selected}
		\item{\code{procId}: }{id of selected process or -1 if nothing is selected}
		\item{\\code{runIdList}: }{ilist of possible process run ids}
}

List of tags.
}
\description{
Landing Page

landingPage
}
