% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingpage_module.R
\name{landingPageUI}
\alias{landingPageUI}
\alias{landingPage}
\title{landingPage}
\usage{
landingPageUI(id)

landingPage(input, output, session, logMessage = message,
  active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{landingPage()}, list of reactives:
\itemize{
		\item{\code{anaid}: }{id of selected analysis or -1 if nothing is selected}
		\item{\code{modelid}: }{id of selected model or -1 if nothing is selected}
}
}
\description{
landingPage

Landing Page
}
