% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageheader_module.R
\name{landingPageButtonUpdate}
\alias{landingPageButtonUpdate}
\title{Landing Page Access Control}
\usage{
landingPageButtonUpdate(session, dbSettings, logMessage = message)
}
\arguments{
\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. [flamingoDB()].}

\item{logMessage}{function that will be passed info messages.}
}
\description{
Disable/Enable menu buttons based on permissions in database.
}
