% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R
\name{landingPageButtonUpdate}
\alias{landingPageButtonUpdate}
\title{Landing Page Access Control}
\usage{
landingPageButtonUpdate(session, dbSettings, logMessage = message)
}
\arguments{
\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{logMessage}{function that will be passed info messages}
}
\description{
Disable/Enable menu buttons based on permissions in database
}
