% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageheader_module.R
\name{landingPageButtonUpdate}
\alias{landingPageButtonUpdate}
\title{Landing Page Access Control}
\usage{
landingPageButtonUpdate(session, dbSettings, logMessage = message)
}
\arguments{
\item{logMessage}{function that will be passed info messages.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}
}
\description{
Disable/Enable menu buttons based on permissions in database.
}
