% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R
\name{landingPageButtonUpdate}
\alias{landingPageButtonUpdate}
\title{Landing Page Access Control}
\usage{
landingPageButtonUpdate(session, dbSettings, logMessage = message)
}
\arguments{
\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{logMessage}{function that will be passed info messages.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}
}
\description{
Disable/Enable menu buttons based on permissions in database.
}
