% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{loadProgrammeData}
\alias{loadProgrammeData}
\title{Call BFE webservice to load programme data}
\usage{
loadProgrammeData(apiSettings, progId, logWarning = function(w) warning,
  ...)
}
\arguments{
\item{apiSettings}{settings object as returned by \link{flamingoServer}}

\item{progId}{programme identifier}

\item{logWarning}{warning message callback}

\item{...}{additional arguments to \link{httr::GET}}
}
\value{
http status category warn_for_status
}
\description{
A function to load Programme data.
}
