% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{loadProgrammeModel}
\alias{loadProgrammeModel}
\title{Call BFE webservice to load programme data}
\usage{
loadProgrammeModel(apiSettings, progOasisId, logWarning = function(w) warning,
  ...)
}
\arguments{
\item{apiSettings}{settings object as returned by \link{flamingoServer}}

\item{progOasisId}{oasis programme identifier}

\item{logWarning}{warning message callback}

\item{...}{additional arguments to \link{httr::GET}}
}
\value{
http status category
}
\description{
A function to load Programme data.
}
