% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_server.R, R/login_ui.R
\name{loginDialog}
\alias{loginDialog}
\alias{loginDialogUI}
\title{loginDialog}
\usage{
loginDialog(input, output, session, dbSettings, logout,
  logMessage = message, logError = logMessage)

loginDialogUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{logout}{Reactive yielding logout signal.}

\item{logMessage}{function that will be passed info messages.}

\item{logError}{function that will be passed error messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
List of reactive expressions:
\itemize{
		\item{\code{userId}: }{yielding an user id if login has been completed
					successfully and \link{FLAMINGO_GUEST_ID} otherwise}
		\item{\code{logout}: }{reactive yielding logout button signal}
}.

List of tags.
}
\description{
Server logic to login an user.

UI/View to login an user.
}
