% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_server.R, R/login_ui.R
\name{loginDialog}
\alias{loginDialog}
\alias{loginDialogUI}
\title{Login Dialog Module}
\usage{
loginDialog(input, output, session, dbSettings, logout,
  logMessage = message, logError = logMessage)

loginDialogUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{logout}{reactive yielding logout signal}

\item{id}{account id}
}
\value{
list of reactive expressions
\itemize{
		\item{\code{userId}: }{yielding an user id if login has been completed
					successfully and \link{FLAMINGO_GUEST_ID} otherwise}
		\item{\code{logout}: }{reactive yielding logout button signal}
}
}
\description{
Server logic to login an user

UI/View to login an user
}
