% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_server.R, R/model_ui.R
\name{modelSupplierPage}
\alias{modelSupplierPage}
\alias{modelSupplierPageUI}
\title{modelSupplierPage}
\usage{
modelSupplierPage(input, output, session, dbSettings,
  logMessage = message, active = reactive(TRUE))

modelSupplierPageUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

Empty list.

List of tags.
}
\description{
Server logic for the model supplier page.

UI/View for the model supplier page.
}
