% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_server.R, R/model_ui.R
\name{modelSupplierPage}
\alias{modelSupplierPage}
\alias{modelSupplierPageUI}
\title{modelSupplierPage}
\usage{
modelSupplierPage(input, output, session, dbSettings,
  logMessage = message, active = reactive(TRUE))

modelSupplierPageUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{dbSettings}{Setting object as returned by e.g. [flamingoDB()].}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

Empty list.

List of tags.
}
\description{
Server logic for the model supplier page.

UI/View for the model supplier page.
}
