% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_server.R, R/model_ui.R
\name{modelSupplierPage}
\alias{modelSupplierPage}
\alias{modelSupplierPageUI}
\title{Module For The Model Supplier Page}
\usage{
modelSupplierPage(input, output, session, dbSettings,
  logMessage = message, active = reactive(TRUE))

modelSupplierPageUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{logMessage}{function that will be passed info messages}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{shiny module id}
}
\value{
empty list
}
\description{
Server logic for the model supplier page

UI/View for the model supplier page
}
