% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ui.R
\name{modelSupplierPageUI}
\alias{modelSupplierPageUI}
\title{modelSupplierPageUI}
\usage{
modelSupplierPageUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
List of tags.
}
\description{
UI/View for the model supplier page.
}
