% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfiles_module.R
\name{outputfilesUI}
\alias{outputfilesUI}
\alias{outputfiles}
\title{outputfilesUI}
\usage{
outputfilesUI(id)

outputfiles(input, output, session, dbSettings, apiSettings,
  filesListDatatoview, active, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{filesListDatatoview}{dataframe of files to view}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{user}{reactive expression yielding user}
}
\value{
List of tags.
}
\description{
UI/View for output files of an analysis.

Server logic for output files of an analysis.
}
