% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfiles_module.R
\name{outputfilesUI}
\alias{outputfilesUI}
\alias{outputfiles}
\title{outputfilesUI}
\usage{
outputfilesUI(id)

outputfiles(input, output, session,
  tbl_filesListDataana = reactive(NULL), anaId = reactive(""),
  portfolioId = reactive(""), active = reactive(TRUE),
  counter = reactive(NULL))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{tbl_filesListDataana}{Tbl of output files to view.}

\item{anaId}{Id of analysis.}

\item{portfolioId}{Id of portfolio associated to the analysis.}

\item{active}{reactive expression whether the module state should be updated.}

\item{counter}{Reactive value to trigger inputs download.}
}
\value{
List of tags.
}
\description{
UI/View for output files of an analysis.

Server logic for output files of an analysis.
}
