% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfiles_module.R
\name{outputfiles}
\alias{outputfiles}
\title{outputfiles_server
Run outputfiles Server}
\usage{
outputfiles(input, output, session, dbSettings, apiSettings, userId,
  filesListDatatoview, active, logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags
}
\description{
output files of a Run
}
