% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{outputplots}
\alias{outputplots}
\title{outputplots_server
Run outputplots Server}
\usage{
outputplots(input, output, session, dbSettings, apiSettings, userId,
  selectRunID, n_panels, filesListData = reactive(NULL), active,
  logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags
}
\description{
outputplots of a Run
}
