% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{outputplotsUI}
\alias{outputplotsUI}
\alias{outputplots}
\title{outputplotsUI}
\usage{
outputplotsUI(id)

outputplots(input, output, session, selectAnaID, n_panels,
  filesListData = reactive(NULL), active, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.
}
\description{
UI/View for output plots of a run.

Server logic for outputplots of a run.
}
