% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R
\name{pageheader}
\alias{pageheader}
\title{pageheader}
\usage{
pageheader(input, output, session, userId, userName, dbSettings,
  reloadMillis = 10000, logMessage = message, active = reactive(TRUE))
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{userId}{reactive expression yielding user id}

\item{userName}{reactive expression yielding user name}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{logMessage}{function that will be passed info messages}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
list of reactive expressions
\itemize{
		\item{\code{navigate}: }{reactive yielding navigation}
		\item{\code{logout}: }{reactive yielding logout button signal}
}
}
\description{
pageheader
}
