% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageheader_module.R
\name{pageheaderUI}
\alias{pageheaderUI}
\alias{pageheader}
\title{pageheaderUI}
\usage{
pageheaderUI(id)

pageheader(input, output, session, user, reloadMillis = 10000,
  logMessage = message, active = reactive(TRUE))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{user}{reactive expression yielding user name}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{pageheader()}, list of reactives:
\itemize{
		\item{\code{logout}: }{reactive yielding logout button signal}
}.
}
\description{
pageheaderUI

pageheader
}
