% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingPage_server.R
\name{pagestructure}
\alias{pagestructure}
\title{Page Structure}
\usage{
pagestructure(input, output, session, userId, userName, dbSettings,
  reloadMillis = 10000, logMessage = message,
  active = reactive(TRUE))
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{logMessage}{Function that will be passed info messages.}

\item{active}{Reactive expression whether the module state should be updated.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{pagestructure()}, list of reactives.
}
\description{
Page Structure
}
