% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagestructure_module.R
\name{pagestructureUI}
\alias{pagestructureUI}
\alias{pagestructure}
\title{pagestructureUI}
\usage{
pagestructureUI(id)

pagestructure(input, output, session, logMessage = message,
  active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
Lits of tags

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

collapsed status of panel.
}
\description{
pagestructureUI

Page Structure
}
