% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagestructure_module.R
\name{pagestructureUI}
\alias{pagestructureUI}
\alias{pagestructure}
\title{pagestructureUI}
\usage{
pagestructureUI(id)

pagestructure(input, output, session, reloadMillis = 10000,
  logMessage = message, active = reactive(TRUE))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{logMessage}{function that will be passed info messages.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}
}
\value{
Lits of tags

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

collapsed status of panel.
}
\description{
pagestructureUI

Page Structure
}
