% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelAnalysisDetails_module.R
\name{panelAnalysisDetailsUI}
\alias{panelAnalysisDetailsUI}
\alias{panelAnalysisDetails}
\title{panelAnalysisDetailsUI}
\usage{
panelAnalysisDetailsUI(id)

panelAnalysisDetails(input, output, session, analysisID, tbl_filesListData,
  param, file_column, folderpath, reload_generated, reload_uploaded)
}
\arguments{
\item{id}{module id}

\item{analysisID}{Selected analysis ID.}

\item{tbl_filesListData}{Dataframe of the output files}

\item{param}{analysisID}

\item{file_column}{Name of the column containing filename. Default "files"}

\item{folderpath}{path to files. Can be "_output/output/" or "_inputs/"; default output path.}

\item{reload_generated}{Imports function to reload Generated Inputs table.}

\item{reload_uploaded}{Imports function to reload Uploaded Inputs table.}

\item{id}{module id}
}
\description{
UI side of function wrapping panel to show analyses details table.

Server side of function wrapping panel to show analyses details table.
}
