% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelAnalysisDetails_module.R
\name{panelAnalysisDetailsUI}
\alias{panelAnalysisDetailsUI}
\alias{panelAnalysisDetails}
\title{panelAnalysisDetailsUI}
\usage{
panelAnalysisDetailsUI(id)

panelAnalysisDetails(input, output, session, analysisID, tbl_filesListData,
  param, file_column, folderpath)
}
\arguments{
\item{id}{module id}

\item{id}{module id}
}
\description{
UI side of function wrapping panel to show analyses details table.

Server side of function wrapping panel to show analyses details table.
}
