% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelAnalysisDetails_module.R
\name{panelAnalysisDetailsUI}
\alias{panelAnalysisDetailsUI}
\alias{panelAnalysisDetails}
\title{panelAnalysisDetailsUI}
\usage{
panelAnalysisDetailsUI(id, button)

panelAnalysisDetails(input, output, session, analysisID, anaName,
  portfolioID)
}
\arguments{
\item{id}{module id}

\item{analysisID}{Selected analysis ID.}

\item{anaName}{Analysis name.}

\item{portfolioID}{Selected portfolio ID.}

\item{tbl_filesListData}{Dataframe of the output files}

\item{reload_generated}{Imports function to reload Generated Inputs table.}

\item{id}{module id}
}
\description{
UI side of function wrapping panel to show analyses details table.

Server side of function wrapping panel to show analyses details table.
}
