% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelDefineIDs}
\alias{panelDefineIDs}
\title{panelDefineIDs}
\usage{
panelDefineIDs(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\value{
List of tags.
}
\description{
Function wrapping panel to define prgramme and model IDs.
}
