% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{panelDefineOutputConfiguration}
\alias{panelDefineOutputConfiguration}
\title{panelDefineOutputConfiguration}
\usage{
panelDefineOutputConfiguration(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping sub-panel to define outputs configuration.
}
