% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{panelDefineOutputsDetails}
\alias{panelDefineOutputsDetails}
\title{panelDefineOutputsDetails}
\usage{
panelDefineOutputsDetails(id)
}
\arguments{
\item{id}{Namespace ID for the module.}
}
\description{
Function wrapping sub-panel to define outputs details.
}
