% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_choosePortfolio_module.R
\name{panelDefineProgramme}
\alias{panelDefineProgramme}
\title{panelDefineProgramme}
\usage{
panelDefineProgramme(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to create/amend programme.
}
