% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_choosePortfolio.R
\name{panelLinkFiles}
\alias{panelLinkFiles}
\title{panelLinkFiles}
\usage{
panelLinkFiles(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to link files to a programme.
}
