% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{panelLinkFiles}
\alias{panelLinkFiles}
\title{Function wrapping panel to link files to a programme}
\usage{
panelLinkFiles(id)
}
\description{
Function wrapping panel to link files to a programme
}
