% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelModelDetails}
\alias{panelModelDetails}
\title{Function wrapping panel to show programme model details table}
\usage{
panelModelDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show programme model details table
}
