% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel.R
\name{panelModelDetails}
\alias{panelModelDetails}
\title{panelModelDetails}
\usage{
panelModelDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show details of programme table.
}
