% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{panelOutput}
\alias{panelOutput}
\title{panelOutput}
\usage{
panelOutput(id)
}
\arguments{
\item{id}{module id}
}
\value{
List of tags.
}
\description{
panelOutput
}
\details{
Function wrapping panel to define prgramme and model IDs
}
