% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{panelOutput}
\alias{panelOutput}
\title{panelSummaryTableModuleUI}
\usage{
panelOutput(id)
}
\arguments{
\item{id}{module id}
}
\description{
panelSummaryTableModuleUI
}
\details{
Function wrapping panel to define prgramme and model IDs
}
