% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{panelOutputModule}
\alias{panelOutputModule}
\title{Module for Output Panel}
\usage{
panelOutputModule(input, output, session, logMessage = message,
  filesListData = reactive(NULL), active)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{logMessage}{function that will be passed info messages}

\item{filesListData}{table of output files for a given runID}

\item{active}{reactive expression whether the module state should be updated.}
}
\value{
reactive value of the title
}
\description{
Server logic to show graphical output such as plots
}
