% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{panelOutputModuleUI}
\alias{panelOutputModuleUI}
\alias{panelOutputModule}
\title{panelOutputModuleUI}
\usage{
panelOutputModuleUI(id)

panelOutputModule(input, output, session, logMessage = message, anaID,
  filesListData = reactive(NULL), active)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{logMessage}{function that will be passed info messages.}

\item{anaID}{is selectAnaID}

\item{filesListData}{table of output files for a given anaID}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}
}
\value{
reactive value of the title
}
\description{
Server logic to show graphical output such as plots.
}
