% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R
\name{panelOutputModule}
\alias{panelOutputModule}
\title{Module for Output Panel}
\usage{
panelOutputModule(input, output, session, logMessage = message,
  filesListData, active)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{logMessage}{Function that will be passed info messages.}

\item{filesListData}{table of output files for a given runID}

\item{active}{Reactive expression whether the module state should be updated.}
}
\value{
reactive value of the title
}
\description{
Server logic to show graphical output such as plots
}
