% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{panelOutputModuleUI}
\alias{panelOutputModuleUI}
\title{panelOutputModuleUI}
\usage{
panelOutputModuleUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
List of tags.
}
\description{
panelOutputModuleUI
}
