% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panelProcessRunLogs}
\alias{panelProcessRunLogs}
\title{panelProcessRunLogs}
\usage{
panelProcessRunLogs(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show log table for specific Process Run.
}
