% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{panelProcessRunTable}
\alias{panelProcessRunTable}
\title{panelProcessRunTable}
\usage{
panelProcessRunTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show process run table.
}
