% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelProgrammeDetails}
\alias{panelProgrammeDetails}
\title{Function wrapping panel to show programme details table}
\usage{
panelProgrammeDetails(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show programme details table
}
