% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_ui.R
\name{panelProgrammeModelTable}
\alias{panelProgrammeModelTable}
\title{Function wrapping panel to show created programme model table}
\usage{
panelProgrammeModelTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show created programme model table
}
