% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{panelProgrammeTable}
\alias{panelProgrammeTable}
\title{Function wrapping panel to show created programmes table}
\usage{
panelProgrammeTable(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function wrapping panel to show created programmes table
}
