% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_server.R
\name{panelSummaryTableModule}
\alias{panelSummaryTableModule}
\title{Module for Summary Table Panel}
\usage{
panelSummaryTableModule(input, output, session, dbSettings, apiSettings,
  userId, logMessage = message, selectRunID)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{logMessage}{function that will be passed info messages}

\item{selectRunID}{selected runID}
}
\value{
null
}
\description{
Server logic to show the summary table output
}
