% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R
\name{panelSummaryTableModule}
\alias{panelSummaryTableModule}
\title{panelSummaryTableModule}
\usage{
panelSummaryTableModule(input, output, session, dbSettings, apiSettings,
  userId, logMessage = message, selectRunID)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{logMessage}{function that will be passed info messages.}

\item{selectRunID}{selected runID

#' @return Empty list.}

\item{active}{reactive expression whether the module state should be updated.}
}
\description{
Server logic to show the summary table output
}
