% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{panelSummaryTableModuleUI}
\alias{panelSummaryTableModuleUI}
\title{panelSummaryTableModuleUI}
\usage{
panelSummaryTableModuleUI(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\description{
panelSummaryTableModuleUI
}
