% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{panelSummaryTableModuleUI}
\alias{panelSummaryTableModuleUI}
\title{panelSummaryTableModuleUI}
\usage{
panelSummaryTableModuleUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
panelSummaryTableModuleUI
}
