% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_ui.R
\name{panelViewOutputFiles}
\alias{panelViewOutputFiles}
\title{panelViewOutputFiles}
\usage{
panelViewOutputFiles(id)
}
\arguments{
\item{id}{Shiny module id.}
}
\description{
panelViewOutputFiles
}
