% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotAEPCurve}
\alias{plotAEPCurve}
\title{plotAEPCurve}
\usage{
plotAEPCurve(AEPData, years = 1000)
}
\arguments{
\item{AEPData}{AEP dataframe.}

\item{years}{Years, integer.}
}
\value{
AEP Curve plot.
}
\description{
Plot AEP Curve.
}
