% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGUL}
\alias{plotGUL}
\title{GUL plot}
\usage{
plotGUL(outputPlotData, interactive = FALSE, ...)
}
\arguments{
\item{outputPlotData}{plotly output.}

\item{interactive}{Create interactive plot using \code{\link[plotly:plot_ly]{plotly::plot_ly()}}.}

\item{...}{Extra arguments to \code{\link[=plot]{plot()}}.}
}
\value{
Nothing; the interactive plot object if \code{interactive = TRUE}.
}
\description{
Plot GUL using the current device.
}
