% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_details_module.R
\name{portfolio_detailsUI}
\alias{portfolio_detailsUI}
\alias{portfolio_details}
\title{portfolio_detailsUI}
\usage{
portfolio_detailsUI(id)

portfolio_details(input, output, session, refresh_opt = TRUE,
  portfolioID, counter = reactive(NULL), active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{refresh_opt}{Option to hide/show refresh button.}

\item{portfolioID}{selected portfolio ID.}

\item{counter}{Reactive value storing actionButton status.}
}
\value{
List of tags.
}
\description{
UI/View for portfolio details.

Server logic for portfolio details.
}
