% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionBatch_server.R,
%   R/programmeDefinitionBatch_ui.R
\name{programmeDefinitionBatch}
\alias{programmeDefinitionBatch}
\alias{programmeDefinitionBatchUI}
\title{programmeDefinitionBatch_server}
\usage{
programmeDefinitionBatch(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

programmeDefinitionBatchUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

List of tags.
}
\description{
Server logic to define a batch process.

UI/View for the process batch run page.
}
