% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_server.R,
%   R/programmeDefinitionSingle_ui.R
\name{programmeDefinitionSingle}
\alias{programmeDefinitionSingle}
\alias{programmeDefinitionSingleUI}
\title{Single Programme Definition Module}
\usage{
programmeDefinitionSingle(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  preselRunId = reactive(-1), preselProcId = reactive(-1),
  preselPanel = panelsProgrammeWorkflow[1], reloadMillis = 10000)

programmeDefinitionSingleUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{module id}
}
\value{
For \code{programmeDefinitionSingle()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic to define a programme

UI/View for the process run page
}
