% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveConditionalPanel.R
\name{reactiveConditionalPanel}
\alias{reactiveConditionalPanel}
\title{Reactive Conditional Panel}
\usage{
reactiveConditionalPanel(input, output, session, condition)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{condition}{Reactive boolean expression determining the visibility of
the panel.}
}
\value{
smth.
}
\description{
Shiny module handling the visibility of a panel based on a reactive \code{condition}.
}
\details{
In comparison to \code{\link[shiny:conditionalPanel]{shiny::conditionalPanel()}}, the
\emph{reactiveConditionalPanel} module provides a convenient abstraction from the
JavaScript condition details, relying on the reactive \code{condition} instead.
}
\note{
The module relies on the package \pkg{shinyjs}, which means
\code{\link[shinyjs:useShinyjs]{shinyjs::useShinyjs()}} must be included in the app \code{ui} definition.
}
\examples{
if (interactive()) {
  library(shiny)
  panels <- LETTERS[1:10]
  ui <- fluidPage(
    shinyjs::useShinyjs(),
    titlePanel("Reactive panel visibility `conditionalPanel`"),
    sidebarLayout(
      sidebarPanel(sliderInput(
        "active_nr", "Number of active panels",
        min = 0, max = length(panels), step = 1, round = TRUE, value = 0
      )),
      mainPanel(mapply(
        reactiveConditionalPanelUI,
        panels,  lapply(panels, div),
        SIMPLIFY = FALSE
      ))
    )
  )
  server <- function(input, output) {
    lapply(panels, function(x) {
      callModule(
        reactiveConditionalPanel, x,
        reactive(x \%in\% head(panels, input$active_nr))
      )
    })
  }
  shinyApp(ui = ui, server = server)
}
}
\seealso{
Module \emph{\link{reactiveConditionalPanels}}, providing similar
functionality for multiple panels.
}
