% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveConditionalPanels.R
\name{reactiveConditionalPanels}
\alias{reactiveConditionalPanels}
\alias{reactiveConditionalPanelsUI}
\title{reactiveConditionalPanels}
\usage{
reactiveConditionalPanels(input, output, session, visible)

reactiveConditionalPanelsUI(id, panels)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{visible}{Reactive string expression determining the visibility of
panels.}

\item{id}{Namespace ID for the module.}

\item{panels}{Named list of UI content for a set of panels. Only the panel
whose name matches the \code{visible} reactive input to the module server
function is shown.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{logMessage}{function that will be passed info messages.}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{id}{module id}
}
\value{
smth.

Panels.
}
\description{
Shiny module handling the visibility of a set of panels depending on a
reactive input.
}
\details{
In comparison to \code{\link[shiny:conditionalPanel]{shiny::conditionalPanel()}}, the
\emph{reactiveConditionalPanels} module provides a convenient abstraction from the
JavaScript condition details and only relies on the reactive input
\code{visible} to determine which panels are visible.
}
\note{
The module relies on the package \pkg{shinyjs}, which means
\code{\link[shinyjs:useShinyjs]{shinyjs::useShinyjs()}} must be included in the app \code{ui} definition.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    shinyjs::useShinyjs(),
    titlePanel("Multiple reactive panel visibility via `reactiveConditionalPanels`"),
    sidebarLayout(
      sidebarPanel(
        radioButtons(
          "main_visible",
          "Active main panel",
          LETTERS[1:10]
        )
      ),
      mainPanel(
        reactiveConditionalPanelsUI(
          "main",
          c(
            mapply(SIMPLIFY = FALSE, textInput,
                   LETTERS[1:5], paste("Main Panel", LETTERS[1:5])),
            list(A = textInput("A-nother", "A-nother Main Panel"))
          )
        )
      )
    )
  )
  server <- function(input, output) {
    callModule(reactiveConditionalPanels, "main", reactive(input$main_visible))
  }
  shinyApp(ui = ui, server = server)
}
}
\seealso{
Module \emph{\link{reactiveConditionalPanel}}, providing similar functionality
for a single panel.
}
