% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{renderflamingoPanelHeading}
\alias{renderflamingoPanelHeading}
\title{renderflamingoPanelHeading}
\usage{
renderflamingoPanelHeading(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{Expression.}

\item{env}{Environment.}

\item{...}{Additional parameters.}
}
\value{
Flamingo heading.
}
\description{
renderflamingoPanelHeading
}
