% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoPanel.R
\name{renderFlamingoPanelHeading}
\alias{renderFlamingoPanelHeading}
\title{Render Heading UI}
\usage{
renderFlamingoPanelHeading(expr, env = parent.frame(), ...)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{...}{Further arguments passed to \code{\link[shiny:renderUI]{shiny::renderUI()}}.}
}
\description{
Render reactive panel heading content dynamically based on server logic.
}
\seealso{
\code{\link[=flamingoPanelHeadingOutput]{flamingoPanelHeadingOutput()}}
}
