% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_server.R
\name{replaceWithIcons}
\alias{replaceWithIcons}
\title{Function to replace status with icons in table}
\usage{
replaceWithIcons(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Function to replace status with icons in table
}
