% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeDefinitionSingle_server.R
\name{replaceWithIcons}
\alias{replaceWithIcons}
\title{replaceWithIcons}
\usage{
replaceWithIcons(df)
}
\arguments{
\item{df}{\code{data.frame}.}
}
\description{
Function to replace status with icons in table.
}
