% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{return_df}
\alias{return_df}
\title{Return dataframe from API response}
\usage{
return_df(api_query, api_param = "")
}
\arguments{
\item{api_query}{Function representing the API query.}

\item{api_param}{Parameter for the api_query.}
}
\value{
Dataframe of the API response.
}
\description{
Returns a dataframe of the API response.
}
