% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{return_tbl_modelData}
\alias{return_tbl_modelData}
\title{Return Model Data fot DT}
\usage{
return_tbl_modelData(id)
}
\arguments{
\item{id}{the ID for the model.}
}
\value{
dataframe of previously posted model.
}
\description{
Returns a dataframe of model ready for being rendered as a data table
}
