% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_models.R
\name{return_tbl_modelsData}
\alias{return_tbl_modelsData}
\title{Return models data for DT}
\usage{
return_tbl_modelsData(supplier_id = "")
}
\arguments{
\item{supplier_id}{The supplier ID for the model. Default is empty string.}
}
\value{
Dataframe of previously posted models. Default empty string returns all models.
}
\description{
Returns a dataframe of models ready for being rendered as a data table.
}
