% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{return_tbl_portfolioDetails}
\alias{return_tbl_portfolioDetails}
\title{Return portfolio details for DT}
\usage{
return_tbl_portfolioDetails(id)
}
\arguments{
\item{id}{Id of the portfolio.}
}
\value{
Dataframe of details of previously posted portfolio.
}
\description{
Returns a dataframe of portfolio's details ready for being rendered as a data table.
}
