% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_portfolios.R
\name{return_tbl_portfoliosData}
\alias{return_tbl_portfoliosData}
\title{Return portfolios data for DT}
\usage{
return_tbl_portfoliosData(name = "", oasisapi)
}
\arguments{
\item{name}{Name of the portfolio.}

\item{oasisapi}{as stored in session$userData$oasisapi}
}
\value{
Dataframe of previously posted portfolios. Default empty string returns all portfolios.
}
\description{
Returns a dataframe of portfolios ready for being rendered as a data table.
}
