% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlamingo.R
\name{runFlamingo}
\alias{runFlamingo}
\title{runFlamingo}
\usage{
runFlamingo(port = 3838, ...)
}
\arguments{
\item{port}{port to pass to \code{\link[shiny]{runApp}}.}

\item{...}{Additional arguments to \code{\link[shiny]{runApp}}.}
}
\value{
No return value.
}
\description{
Runs the Flamingo Shiny application. Under normal circumstances
this function does not return. The application is run in a temporary
directory and is available under the specified port.
}
