% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{runProcess}
\alias{runProcess}
\title{runProcess}
\usage{
runProcess(apiSettings, runId, logWarning = warning, ...)
}
\arguments{
\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{runId}{Run identifier.}

\item{logWarning}{Warning message callback-}

\item{...}{Further arguments to \link{GET}.}
}
\value{
http status category warn_for_status.
}
\description{
Submit a GET request to the Flamingo API Server to execute
the workflow.
}
