% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{showname}
\alias{showname}
\title{Return element based on length of input}
\usage{
showname(x)
}
\arguments{
\item{x}{List, NULL or string.}
}
\value{
String.
}
\description{
Returns the element name of the input x if x has length > 1, the
  element x itself if it has length = 1, and "Not available" if x has length
  = 0.
}
