% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleProgrammeWorkflowSteps.R
\name{singleProgrammeWorkflowSteps}
\alias{singleProgrammeWorkflowSteps}
\title{singleProgrammeWorkflowSteps}
\usage{
singleProgrammeWorkflowSteps(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{id}{Namespace ID for the module.}
}
\value{
smth.
}
\description{
singleProgrammeWorkflowSteps
}
