% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{step1_chooseProgrammeUI}
\alias{step1_chooseProgrammeUI}
\alias{step1_chooseProgramme}
\title{step1_chooseProgramme UI}
\usage{
step1_chooseProgrammeUI(id)

step1_chooseProgramme(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive(""))
}
\arguments{
\item{id}{module id}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags

For \code{programmeDefinitionSingle()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View for the step1_chooseProgramme

Server logic to step1_chooseProgramme
}
