% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_chooseProgramme.R
\name{step1_chooseProgrammeUI}
\alias{step1_chooseProgrammeUI}
\alias{step1_chooseProgramme}
\title{step1_chooseProgrammeUI}
\usage{
step1_chooseProgrammeUI(id)

step1_chooseProgramme(input, output, session, dbSettings, apiSettings,
  userId, active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive(""))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{currstep}{current selected step.}

\item{selectprogrammeID}{selected programme ID.}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

selectprogrammeID Id of selected programmeID

DPProgData POData model association table.

newstep navigation step
}
\description{
UI/View for the step1_chooseProgramme.

Server logic to step1_chooseProgramme.
}
