% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel.R
\name{step2_chooseModelUI}
\alias{step2_chooseModelUI}
\alias{step2_chooseModel}
\title{step2_chooseModel UI}
\usage{
step2_chooseModelUI(id)

step2_chooseModel(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive({     "" }),
  selectprogOasisID = reactive({     "" }), progName = reactive({    
  "" }), progStatus = reactive({     "" }),
  DPProgData = reactive(NULL))
}
\arguments{
\item{id}{Shiny module id.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

For \code{programmeDefinitionSingle()}, list of reactives.
}
\description{
UI/View for the step2_chooseModel.

Server logic to step2_chooseModel.
}
