% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_chooseModel.R
\name{step2_chooseModelUI}
\alias{step2_chooseModelUI}
\alias{step2_chooseModel}
\title{step2_chooseModel UI}
\usage{
step2_chooseModelUI(id)

step2_chooseModel(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive({     "" }),
  selectprogOasisID = reactive({     "" }), progName = reactive({    
  "" }), progStatus = reactive({     "" }),
  DPProgData = reactive(NULL))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{currstep}{current selected step.}

\item{selectprogrammeID}{selected programme ID.}

\item{selectprogOasisID}{selected ProgOasis ID.}

\item{progName}{Name of selected programme.}

\item{progStatus}{Status of selected programme.}

\item{DPProgData}{Dataframe of programmes.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

selectprogOasisID Id of selected progOasis.

POData model association table.

newstep navigation step
}
\description{
UI/View for the step2_chooseModel.

Server logic to step2_chooseModel.
}
