% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput.R
\name{step3_configureOutputUI}
\alias{step3_configureOutputUI}
\alias{step3_configureOutput}
\title{step3_configureOutputUI}
\usage{
step3_configureOutputUI(id)

step3_configureOutput(input, output, session, dbSettings, apiSettings,
  active = reactive(TRUE), logMessage = message,
  currstep = reactive(-1), selectprogrammeID = reactive(""),
  selectprogOasisID = reactive(""), progOasisName = reactive(""),
  progOasisStatus = reactive(""))
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{currstep}{current selected step.}

\item{selectprogrammeID}{selected programme ID.}

\item{selectprogOasisID}{selected ProgOasis ID.}

\item{progOasisName}{Name of selected progOasis.}

\item{progOasisStatus}{Status of selected progOasis.}

\item{id}{Namespace ID for the module.}

\item{user}{reactive expression yielding user}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

prrunid id of selected run.
}
\description{
UI/View for the step3_configureOutput.

Server logic to step3_configureOutput.
}
