% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_configureOutput_module.R
\name{step3_configureOutputUI}
\alias{step3_configureOutputUI}
\alias{step3_configureOutput}
\title{step3_configureOutputUI}
\usage{
step3_configureOutputUI(id)

step3_configureOutput(input, output, session, active = reactive(TRUE),
  logMessage = message, currstep = reactive(-1),
  portfolioID = reactive(""), analysisID = reactive(""))
}
\arguments{
\item{id}{module id}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{currstep}{current selected step.}

\item{portfolioID}{selected portfolio ID.}

\item{analysisID}{selected analysis ID}

\item{id}{Namespace ID for the module.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

anaID id of selected run.
}
\description{
UI/View for the step3_configureOutput.

Server logic to step3_configureOutput.
}
