% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystatus_module.R
\name{summarystatusUI}
\alias{summarystatusUI}
\alias{summarystatus}
\title{summarystatusUI}
\usage{
summarystatusUI(id)

summarystatus(input, output, session, analysisID, counter = NULL,
  active = reactive(TRUE))
}
\arguments{
\item{id}{Namespace ID for the module.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{analysisID}{Selected analysis id.}

\item{counter}{Reactive value storing actionButton status.}
}
\value{
List of tags.
}
\description{
UI/View for summary status of files of an analysis.

Server logic for summary status of files of an analysis.
}
