% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{summarytab}
\alias{summarytab}
\title{summarytab_server
Run Summary Server}
\usage{
summarytab(input, output, session, dbSettings, apiSettings, userId,
  selectRunID = reactive(""), active, logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
list of tags
}
\description{
Summary elements of a Run
}
