% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{summarytabUI}
\alias{summarytabUI}
\alias{summarytab}
\title{summarytab_ui}
\usage{
summarytabUI(id)

summarytab(input, output, session, selectAnaID1 = reactive(""),
  selectAnaID2 = reactive(""), portfolioID1 = reactive(""),
  portfolioID2 = reactive(""), compare = FALSE, active,
  logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}
}
\value{
list of tags

list of tags
}
\description{
UI/View of summary elements of an analysis.

Server logic of summary elements of an analysis.
}
\details{
Analysis Summary Server
}
