% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_module.R
\name{summarytabUI}
\alias{summarytabUI}
\alias{summarytab}
\title{summarytab_ui}
\usage{
summarytabUI(id)

summarytab(input, output, session, dbSettings, apiSettings, userId,
  selectRunID1 = reactive(""), selectRunID2 = reactive(""),
  compare = FALSE, active, logMessage = message)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}
}
\value{
list of tags

list of tags
}
\description{
UI/View of summary elements of a run.

Server logic of summary elements of a run.
}
\details{
Run Summary Server
}
