% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplots_module.R
\name{outputplotsUI}
\alias{outputplotsUI}
\title{outputplots_ui
Run outputplots UI}
\usage{
outputplotsUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
list of tags
}
\description{
Output plots of a Run
}
