% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoServer.R
\name{testFlamingoServer}
\alias{testFlamingoServer}
\title{testFlamingoServer}
\usage{
testFlamingoServer(apiSettings, ...)
}
\arguments{
\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{...}{further arguments to \link{GET}.}
}
\value{
\code{200} if the health check succeeds; otherwise an exception is
raised.
}
\description{
Performs a health check on the Flamingo API Server.
}
