% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{updateModelResource}
\alias{updateModelResource}
\title{Update Model Resource}
\usage{
updateModelResource(dbSettings, modresid, modresname, restype, oasissys,
  modelid, modresval)
}
\arguments{
\item{dbSettings}{setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}}

\item{modresid}{model resource id}

\item{modresname}{\code{character()}; name of the model resource}

\item{restype}{resource type}

\item{oasissys}{TODO document}

\item{modelid}{model id}

\item{modresval}{TODO document}
}
\value{
model resource id
}
\description{
Update Model Resource
}
