% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flamingoDB.R
\name{updateModelResource}
\alias{updateModelResource}
\title{updateModelResource}
\usage{
updateModelResource(dbSettings, modresid, modresname, restype, oasissys,
  modelid, modresval)
}
\arguments{
\item{dbSettings}{as returned from \link{flamingoDB}}

\item{modresid}{Model resource id.}

\item{modresname}{\code{character()}; name of the model resource.}

\item{restype}{Resource type.}

\item{oasissys}{TODO document.}

\item{modelid}{Model id.}

\item{modresval}{TODO document.}
}
\value{
Model resource id.
}
\description{
Update Model Resource.
}
