% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUserAdmin_server.R,
%   R/defineUserAdmin_ui.R
\name{userAdminDefinition}
\alias{userAdminDefinition}
\alias{userAdminDefinitionUI}
\title{User Admin Definition Module}
\usage{
userAdminDefinition(input, output, session, dbSettings, user,
  active = reactive(TRUE))

userAdminDefinitionUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{user}{reactive yielding user (the logged in user)}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{shiny module id}
}
\value{
For \code{userAdminDefinition()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic for accessing the Company User List for Flamingo
in association with OASIS LMF

UI/View for accessing the Company User List for Flamingo
in association with OASIS LMF
}
