% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUserAdmin_server.R,
%   R/defineUserAdmin_ui.R
\name{userAdminDefinition}
\alias{userAdminDefinition}
\alias{userAdminDefinitionUI}
\title{userAdminDefinition}
\usage{
userAdminDefinition(input, output, session, dbSettings, user,
  active = reactive(TRUE))

userAdminDefinitionUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{id}{Namespace ID for the module.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

List of tags.
}
\description{
Server logic for accessing the Company User List for Flamingo
in association with OASIS LMF.

UI/View for accessing the Company User List for Flamingo
in association with OASIS LMF.
}
