% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModule}
\alias{ViewFilesModule}
\title{ViewFilesModule}
\usage{
ViewFilesModule(input, output, session, logMessage = message,
  filesListData, includemrows = FALSE, includechkbox = FALSE)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{logMessage}{Function that will be passed info messages.}

\item{filesListData}{Table of output files for a given runID.}
}
\value{
List of reactives.
}
\description{
Server logic to view files.
}
