% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModuleUI}
\alias{ViewFilesModuleUI}
\alias{ViewFilesModule}
\title{ViewFilesModuleUI}
\usage{
ViewFilesModuleUI(id, includechkbox = FALSE)

ViewFilesModule(input, output, session, dbSettings, logMessage = message,
  filesListData, includechkbox = FALSE)
}
\arguments{
\item{id}{module id}

\item{includechkbox}{logical indicating the presence of checkboxes. Default FALSE.}

\item{input, output, session}{Shiny input, output, and session objects.}

\item{dbSettings}{Setting object as returned by e.g. [flamingoDB()].}

\item{logMessage}{function that will be passed info messages.}

\item{filesListData}{Table of output files for a given anaID.}

\item{id}{Namespace ID for the module.}

\item{includechkbox}{logical indicating the presence of checkboxes. Default FALSE.}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View to view  files.

Server logic to view files.
}
