% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModuleUI}
\alias{ViewFilesModuleUI}
\alias{ViewFilesModule}
\title{ViewFilesModuleUI}
\usage{
ViewFilesModuleUI(id, includechkbox = FALSE)

ViewFilesModule(input, output, session, logMessage = message,
  filesListData, includemrows = FALSE, includechkbox = FALSE)
}
\arguments{
\item{id}{module id}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{logMessage}{function that will be passed info messages.}

\item{filesListData}{Table of output files for a given runID.}

\item{id}{Namespace ID for the module.}

\item{active}{reactive expression whether the module state should be updated.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}
}
\value{
List of tags.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
UI/View to view  files.

Server logic to view files.
}
