% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewFiles.R
\name{ViewFilesModuleUI}
\alias{ViewFilesModuleUI}
\title{ViewFilesModuleUI}
\usage{
ViewFilesModuleUI(id, includechkbox = FALSE)
}
\arguments{
\item{id}{Shiny module id.}
}
\value{
List of tags.
}
\description{
UI logic to view  files.
}
