% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_server.R,
%   R/visualizationBBR_ui.R
\name{visualizationBBR}
\alias{visualizationBBR}
\alias{visualizationBBRUI}
\title{visualizationBBR}
\usage{
visualizationBBR(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

visualizationBBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{Setting object as returned by e.g. \code{\link[=flamingoDB]{flamingoDB()}}.}

\item{apiSettings}{Settings object as returned by \code{\link[=flamingoServer]{flamingoServer()}}.}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{Reactive expression whether the module state should be updated.}

\item{logMessage}{Function that will be passed info messages.}

\item{reloadMillis}{Amount of time to wait between table updates;
see \link{invalidateLater}.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

List of tags.
}
\description{
Server logic for batchbrowse run page.

UI/View for batchbrowse run page.
}
