% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_server.R,
%   R/visualizationBBR_ui.R
\name{visualizationBBR}
\alias{visualizationBBR}
\alias{visualizationBBRUI}
\title{Batch Browse Definition Module}
\usage{
visualizationBBR(input, output, session, dbSettings, apiSettings,
  active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

visualizationBBRUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{module id}
}
\value{
For \code{visualizationBBR()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic to define a programme

UI/View for batchbrowse run page
}
