% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_server.R,
%   R/visualizationBBR_ui.R
\name{visualizationBBR}
\alias{visualizationBBR}
\alias{visualizationBBRUI}
\title{visualizationBBR}
\usage{
visualizationBBR(input, output, session, dbSettings, apiSettings, user,
  preselAnaId = reactive(-1), processRunId = reactive(-1),
  active = reactive(TRUE), logMessage = message)

visualizationBBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingpage}.}

\item{processRunId}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session.

List of tags.
}
\description{
Server logic for batchbrowse run page.

UI/View for batchbrowse run page.
}
