% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationBBR_server.R,
%   R/visualizationBBR_ui.R
\name{visualizationBBR}
\alias{visualizationBBR}
\alias{visualizationBBRUI}
\title{visualizationBBR}
\usage{
visualizationBBR(input, output, session, dbSettings, apiSettings, userId,
  runIdList = reactive(c(-1)), preselRunId = reactive(-1),
  processRunId = reactive(-1), active = reactive(TRUE),
  logMessage = message)

visualizationBBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the process session.

List of tags.
}
\description{
Server logic for batchbrowse run page.

UI/View for batchbrowse run page.
}
