% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R
\name{visualizationCBR}
\alias{visualizationCBR}
\title{Compare Runs Definition Module}
\usage{
visualizationCBR(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), runIdList = reactive(c(-1)),
  preselRunId = reactive(-1), processRunId = reactive(-1),
  logMessage = message)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}
}
\value{
For \code{visualizationCBR()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.
}
\description{
Server logic to compare runs
}
