% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R,
%   R/visualizationCBR_ui.R
\name{visualizationCBR}
\alias{visualizationCBR}
\alias{visualizationCBRUI}
\title{Compare Runs Definition Module Server ---------------------------------------
visualizationCBR}
\usage{
visualizationCBR(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), runIdList = reactive(c(-1)),
  preselRunId = reactive(-1), processRunId = reactive(-1),
  logMessage = message)

visualizationCBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{active}{reactive expression whether the module state should be updated.}

\item{runIdList}{List of runs and their status.}

\item{preselRunId}{Reactive string expression for reselected run id from landingpage.}

\item{processRunId}{Reactive string expression for reselected run id from defineProgramme.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the process session.

List of tags.
}
\description{
Server logic for comparing runs run page.

UI/View for comparing runs run page.
}
