% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R,
%   R/visualizationCBR_ui.R
\name{visualizationCBR}
\alias{visualizationCBR}
\alias{visualizationCBRUI}
\title{Compare Runs Definition Module}
\usage{
visualizationCBR(input, output, session, dbSettings, apiSettings, userId,
  active = reactive(TRUE), logMessage = message,
  reloadMillis = 10000)

visualizationCBRUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{module id}
}
\value{
For \code{visualizationCBR()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic to compare runs

UI/View for compare runs run page
}
