% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R,
%   R/visualizationCBR_ui.R
\name{visualizationCBR}
\alias{visualizationCBR}
\alias{visualizationCBRUI}
\title{Compare Analyses Definition Module Server ---------------------------------------
visualizationCBR}
\usage{
visualizationCBR(input, output, session, active = reactive(TRUE),
  preselAnaId = reactive(-1), anaID = reactive(-1))

visualizationCBRUI(id)
}
\arguments{
\item{input, output, session}{Shiny input, output, and session objects.}

\item{active}{reactive expression whether the module state should be updated.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingPage}.}

\item{anaID}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{id}{Namespace ID for the module.}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session.

List of tags.
}
\description{
Server logic for comparing runs analyses page.

UI/View for comparing runs run page.
}
