% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationCBR_server.R,
%   R/visualizationCBR_ui.R
\name{visualizationCBR}
\alias{visualizationCBR}
\alias{visualizationCBRUI}
\title{Compare Analyses Definition Module Server ---------------------------------------
visualizationCBR}
\usage{
visualizationCBR(input, output, session, dbSettings, apiSettings, user,
  active = reactive(TRUE), preselAnaId = reactive(-1),
  processRunId = reactive(-1), logMessage = message)

visualizationCBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{user}{reactive expression yielding user}

\item{active}{reactive expression whether the module state should be updated.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingpage}.}

\item{processRunId}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{anaIdList}{List of analyses and their status.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session.

List of tags.
}
\description{
Server logic for comparing runs analyses page.

UI/View for comparing runs run page.
}
