% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R,
%   R/visualizationSBR_ui.R
\name{visualizationSBR}
\alias{visualizationSBR}
\alias{visualizationSBRUI}
\title{Single Browse Definition Module}
\usage{
visualizationSBR(input, output, session, dbSettings, apiSettings, userId,
  runIdList, preselRunId, processRunId, active = reactive(TRUE),
  logMessage = message, reloadMillis = 10000)

visualizationSBRUI(id)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater};}

\item{id}{module id}
}
\value{
For \code{visualizationSBR()}, list of reactives.

The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

list of tags
}
\description{
Server logic to define a programme

UI/View for single browse run page
}
