% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R
\name{visualizationSBR}
\alias{visualizationSBR}
\title{visualizationSBR Module}
\usage{
visualizationSBR(input, output, session, dbSettings, apiSettings,
  runIdList = reactive(c(-1)), preselRunId = reactive(-1),
  processRunId = reactive(-1), active = reactive(TRUE),
  logMessage = message, reloadMillis = 10000)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{runIdList}{list of runs and their status}

\item{preselRunId}{reactive string expression for reselected run id from landingpage}

\item{processRunId}{reactive string expression for reselected run id from defineProgramme}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages}

\item{reloadMillis}{amount of time to wait between table updates;
see \link{invalidateLater}}
}
\value{
list of reactives:
}
\description{
Server logic for viewing results of a single run
}
