% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R,
%   R/visualizationSBR_ui.R
\name{visualizationSBR}
\alias{visualizationSBR}
\alias{visualizationSBRUI}
\title{visualizationSBR}
\usage{
visualizationSBR(input, output, session, dbSettings, apiSettings, userId,
  runIdList = reactive(c(-1)), preselRunId = reactive(-1),
  processRunId = reactive(-1), active = reactive(TRUE),
  logMessage = message, reloadMillis = 10000)

visualizationSBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbSettings}{as returned from \link{flamingoDB}}

\item{apiSettings}{as returned from \link{flamingoServer}}

\item{userId}{user id as returned by \code{\link[=flamingoDBLogin]{flamingoDBLogin()}}}

\item{runIdList}{List of runs and their status.}

\item{preselRunId}{reactive string expression for reselected run id from landingpage.}

\item{processRunId}{reactive string expression for reselected run id from defineProgramme.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the process session

List of tags.
}
\description{
Server logic for viewing results of a single run.

UI/View for viewing results of a single run.
}
