% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationSBR_server.R,
%   R/visualizationSBR_ui.R
\name{visualizationSBR}
\alias{visualizationSBR}
\alias{visualizationSBRUI}
\title{visualizationSBR}
\usage{
visualizationSBR(input, output, session, dbSettings, apiSettings, user,
  anaIdList = reactive(c(-1)), preselAnaId = reactive(-1),
  anaID = reactive(-1), active = reactive(TRUE),
  logMessage = message)

visualizationSBRUI(id)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{anaIdList}{List of analyses and their status.}

\item{preselAnaId}{reactive string expression for reselected analysis id from \link{landingpage}.}

\item{anaID}{reactive string expression for reselected run id from \link{step3_configureOutput}.}

\item{active}{reactive expression whether the module state should be updated.}

\item{logMessage}{function that will be passed info messages.}

\item{id}{Namespace ID for the module.}

\item{id}{module id}
}
\value{
The returned list also includes reactive navigation state
constructed via \code{\link[=outputNavigation]{outputNavigation()}}.

preselPanel panel to show in the model session

List of tags.
}
\description{
Server logic for viewing results of a single analysis.

UI/View for viewing results of a single run.
}
